### 2022-08-20 #4690 @arfon
@bwohlberg – please fix the whitespace in [your paper](https://github.com/lanl/scico/blob/joss-paper/docs/joss-paper/paper.md) to match our [example](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) – YAML is whitespace sensitive.

Also, I'm not sure that adding those extra packages in the `header-includes` section will work here, but we can try :-)
