### 2022-08-20 #4690 @bwohlberg
> @bwohlberg – please fix the whitespace in [your paper](https://github.com/lanl/scico/blob/joss-paper/docs/joss-paper/paper.md) to match our [example](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) – YAML is whitespace sensitive.

Sorry about that. We fixed the header a while back, but somehow the bad indentation seems to have made its way back in. It's fixed now.

> Also, I'm not sure that adding those extra packages in the `header-includes` section will work here, but we can try :-)

It worked when we built the paper in the joss github workflow or using the docker image.
