### 2022-08-22 #4690 @arfon
@bwohlberg – thanks for fixing up the paper. While we look for an editor to pick up this submission could you please take a look at substantially reducing the size of your paper please? As a guideline, JOSS papers should be between 250 -- 1000 words (yours is upwards of 3000 currently).
