### 2022-08-23 9ed29b1
Move definitions of Operator and LinearOperator base classes
