+++ scico/operator/_operator.py
@@ -1,0 +1,368 @@
+# Copyright (C) 2020-2022 by SCICO Developers
+# All rights reserved. BSD 3-clause License.
+# This file is part of the SCICO package. Details of the copyright and
+# user license can be found in the 'LICENSE' file distributed with the
+# package.
+
+"""Operator base class."""
+
+
+# Needed to annotate a class method that returns the encapsulating class;
+# see https://www.python.org/dev/peps/pep-0563/
+from __future__ import annotations
+
+from functools import wraps
+from typing import Callable, Optional, Tuple, Union
+
+import numpy as np
+
+import jax
+from jax.dtypes import result_type
+from jax.interpreters.xla import DeviceArray
+
+import scico.numpy as snp
+from scico.numpy import BlockArray
+from scico.numpy.util import is_nested, shape_to_size
+from scico.typing import BlockShape, DType, JaxArray, Shape
+
+
+def _wrap_mul_div_scalar(func):
+    r"""Wrapper function for defining mul, rmul, and truediv.
+
+    Wrapper function for defining mul, rmul, and truediv between a scalar
+    and an Operator.
+
+    If one of these binary operations are called in the form
+    binop(Operator, other) and 'b' is a scalar, specialized
+    Operator constructors can be called.
+
+    Args:
+        func: should be either .__mul__(), .__rmul__(),
+           or .__truediv__().
+
+    Raises:
+        TypeError: A binop with the form binop(Operator, other) is
+        called and other is not a scalar.
+    """
+
+    @wraps(func)
+    def wrapper(a, b):
+        if np.isscalar(b) or isinstance(b, jax.core.Tracer):
+            return func(a, b)
+
+        raise TypeError(f"Operation {func.__name__} not defined between {type(a)} and {type(b)}")
+
+    return wrapper
+
+
+class Operator:
+    """Generic Operator class."""
+
+    def __repr__(self):
+        return f"""{type(self)}
+shape       : {self.shape}
+matrix_shape : {self.matrix_shape}
+input_dtype : {self.input_dtype}
+output_dtype : {self.output_dtype}
+        """
+
+    # See https://docs.scipy.org/doc/numpy-1.10.1/user/c-info.beyond-basics.html#ndarray.__array_priority__
+    __array_priority__ = 1
+
+    def __init__(
+        self,
+        input_shape: Union[Shape, BlockShape],
+        output_shape: Optional[Union[Shape, BlockShape]] = None,
+        eval_fn: Optional[Callable] = None,
+        input_dtype: DType = np.float32,
+        output_dtype: Optional[DType] = None,
+        jit: bool = False,
+    ):
+        r"""
+        Args:
+            input_shape: Shape of input array.
+            output_shape: Shape of output array.
+                Defaults to ``None``. If ``None``, `output_shape` is
+                determined by evaluating `self.__call__` on an input
+                array of zeros.
+            eval_fn: Function used in evaluating this Operator.
+                Defaults to ``None``. If ``None``, then `self.__call__`
+                must be defined in any derived classes.
+            input_dtype: `dtype` for input argument.
+                Defaults to ``float32``. If `Operator` implements
+                complex-valued operations, this must be ``complex64`` for
+                proper adjoint and gradient calculation.
+            output_dtype: `dtype` for output argument.
+                Defaults to ``None``. If ``None``, `output_shape` is
+                determined by evaluating `self.__call__` on an input
+                array of zeros.
+            jit: If ``True``, call :meth:`Operator.jit()` on this
+                Operator to jit the forward, adjoint, and gram functions.
+                Same as calling :meth:`Operator.jit` after the Operator
+                is created.
+        """
+
+        #: Shape of input array or :class:`.BlockArray`.
+        self.input_shape: Union[Shape, BlockShape]
+
[…]
