### 2022-08-23 a845b2c
Move definition of linops based on functions to new module
