### 2022-08-23 8888e5f
Move misc linop definitions into new module
