+++ scico/linop/_linop.py
@@ -24,15 +24,10 @@ from jax.interpreters.xla import DeviceArray
 import scico.numpy as snp
 from scico._autograd import linear_adjoint
 from scico.numpy import BlockArray
-from scico.numpy.util import (
-    ensure_on_device,
-    indexed_shape,
-    is_complex_dtype,
-    is_nested,
-)
+from scico.numpy.util import is_complex_dtype
 from scico.operator._operator import Operator, _wrap_mul_div_scalar
 from scico.random import randn
-from scico.typing import ArrayIndex, BlockShape, DType, JaxArray, PRNGKey, Shape
+from scico.typing import BlockShape, DType, JaxArray, PRNGKey, Shape
 
 
 def power_iteration(A: LinearOperator, maxiter: int = 100, key: Optional[PRNGKey] = None):
@@ -595,144 +590,3 @@ class ComposedLinearOperator(LinearOperator):
             adj_fn=lambda z: self.B.adj(self.A.adj(z)),
             jit=jit,
         )
-
-
-class Diagonal(LinearOperator):
-    """Diagonal linear operator."""
-
-    def __init__(
-        self,
-        diagonal: Union[JaxArray, BlockArray],
-        input_shape: Optional[Shape] = None,
-        input_dtype: Optional[DType] = None,
-        **kwargs,
-    ):
-        r"""
-        Args:
-            diagonal: Diagonal elements of this linear operator.
-            input_shape:  Shape of input array. By default, equal to
-               `diagonal.shape`, but may also be set to a shape that is
-               broadcast-compatiable with `diagonal.shape`.
-            input_dtype: `dtype` of input argument. The default,
-               ``None``, means `diagonal.dtype`.
-        """
-
-        self.diagonal = ensure_on_device(diagonal)
-
-        if input_shape is None:
-            input_shape = self.diagonal.shape
-
-        if input_dtype is None:
-            input_dtype = self.diagonal.dtype
-
-        if isinstance(diagonal, BlockArray) and is_nested(input_shape):
-            output_shape = (snp.empty(input_shape) * diagonal).shape
-        elif not isinstance(diagonal, BlockArray) and not is_nested(input_shape):
-            output_shape = snp.broadcast_shapes(input_shape, self.diagonal.shape)
-        elif isinstance(diagonal, BlockArray):
-            raise ValueError("`diagonal` was a BlockArray but `input_shape` was not nested.")
-        else:
-            raise ValueError("`diagonal` was a not BlockArray but `input_shape` was nested.")
-
-        super().__init__(
-            input_shape=input_shape,
-            input_dtype=input_dtype,
-            output_shape=output_shape,
-            output_dtype=input_dtype,
-            **kwargs,
-        )
-
-    def _eval(self, x):
-        return x * self.diagonal
-
-    @partial(_wrap_add_sub, op=operator.add)
-    def __add__(self, other):
-        if self.diagonal.shape == other.diagonal.shape:
-            return Diagonal(diagonal=self.diagonal + other.diagonal)
-        raise ValueError(f"Incompatible shapes: {self.shape} != {other.shape}")
-
-    @partial(_wrap_add_sub, op=operator.sub)
-    def __sub__(self, other):
-        if self.diagonal.shape == other.diagonal.shape:
-            return Diagonal(diagonal=self.diagonal - other.diagonal)
-        raise ValueError(f"Incompatible shapes: {self.shape} != {other.shape}")
-
-    @_wrap_mul_div_scalar
-    def __mul__(self, scalar):
-        return Diagonal(diagonal=self.diagonal * scalar)
-
-    @_wrap_mul_div_scalar
-    def __rmul__(self, scalar):
-        return Diagonal(diagonal=self.diagonal * scalar)
-
-    @_wrap_mul_div_scalar
[…]
