### 2022-08-23 1fa45cd
Move utility functions to new module
