### 2022-08-24 64352d3
Resolve mypy errors
