+++ scico/linop/radon_svmbir.py
@@ -22,7 +22,8 @@ import scico.numpy as snp
 from scico.loss import Loss, SquaredL2Loss
 from scico.typing import Array, JaxArray, Shape
 
-from ._linop import Diagonal, Identity, LinearOperator
+from ._linop import LinearOperator
+from ._misc import Diagonal, Identity
 
 try:
     import svmbir
