### 2022-08-24 #4690 @bwohlberg
@arfon: Would you please consider granting an exception on the word count guideline?

We've put considerable effort into writing what we believe is a coherent overview of the purpose and features of the package described by the paper. The "additional" detail addresses the high-level design and usage of the package, and does not involve the sort of API-level details that clearly belong in the project documentation. Cutting it by a factor of three would turn it into a completely different document, and involve removal of a great deal of valuable content.
