### 2022-08-24 #4690 @danielskatz
We (I and @arfon) think this paper needs to be reduced in size. JOSS papers are meant to stand in for the software, not completely document it. Specifically, I suggest that the section "Solving Imaging Inverse Problems in SCICO" could be dramatically reduced, with most of the existing content of this section being moved to a webpage that the shortened section could point to for more details.

While I sympathize with the effort you've put into writing this, the JOSS guidance is quite clear about what a JOSS paper should (and should not) include, and has been in place for years, so I think moving this content to the software repo outside of the paper is the best way to not waste this effort and still meet the JOSS requirements.

We can be somewhat loose in terms of the 1000-word maximum for a paper, but this paper far exceeds that limit and does need to be reduced to get much closer to it.
