### 2022-08-24 #4690 @bwohlberg
Understood. We'll cut it down to the brief abstract-like document that is expected. We had been thinking of JOSS as an alternative venue for the sort of papers that are published in the SciPy Proceedings.
