### 2022-08-24 #4690 @danielskatz
If you want to publish a long paper about the software, you might want to consider options in https://www.software.ac.uk/which-journals-should-i-publish-my-software
