### 2022-08-24 #4690 @danielskatz
We are happy to have your submission, but I did want to offer alternatives. Did you consider [JORS](https://openresearchsoftware.metajnl.com)?  That might meet your needs, though the review process can be somewhat slow
