### 2022-08-24 1559e39
Cut paper to abstract format
