+++ docs/joss-paper/paper.md
@@ -11,22 +11,24 @@ header-includes:
   - \DeclareMathOperator*{\argmin}{arg\,min}
 authors:
   - name: Thilo Balke
-    orcid: 0000-0002-1716-5923
-    affiliation: "1, 2"
+	orcid: 0000-0002-1716-5923
+	affiliation: "1, 2"
   - name: Fernando Davis
-    affiliation: "1, 3"
+	affiliation: "1, 3"
   - name: Cristina Garcia-Cardona
-    orcid: 0000-0002-5641-3491
-    affiliation: 1
+	orcid: 0000-0002-5641-3491
+	affiliation: 1
   - name: Soumendu Majee
-    orcid: 0000-0002-8442-2897
-    affiliation: 4
+	orcid: 0000-0002-8442-2897
+	affiliation: 4
   - name: Michael McCann
-    orcid: 0000-0001-7645-252X
-    affiliation: 1
+	orcid: 0000-0001-7645-252X
+	affiliation: 1
+  - name: Luke Pfister
+	orcid: 0000-0001-7485-5966
   - name: Brendt Wohlberg
-    orcid: 0000-0002-4767-1843
-    affiliation: 1
+	orcid: 0000-0002-4767-1843
+	affiliation: 1
 affiliations:
  - name: Los Alamos National Laboratory
    index: 1
@@ -36,7 +38,7 @@ affiliations:
    index: 3
  - name: Samsung Research America
    index: 4
-date: 20 August 2022
+date: 24 August 2022
 bibliography: paper.bib
 ---
 
@@ -46,7 +48,10 @@ Scientific Computational Imaging Code (SCICO) is a Python package for solving th
 which is used to automatically compute the adjoints of linear operators.
 An example of how to solve a multi-channel tomography problem with SCICO is shown in \autoref{fig:flow_chart}. The SCICO source code is available from [GitHub](https://github.com/lanl/scico), and pre-built packages are available from [PyPI](https://github.com/lanl/scico). It has extensive [online documentation](https://scico.rtfd.io/), including API documentation and usage examples, which can be run online at [Google Colab](https://colab.research.google.com/github/lanl/scico-data/blob/colab/notebooks/index.ipynb) and [binder](https://mybinder.org/v2/gh/lanl/scico-data/binder?labpath=notebooks%2Findex.ipynb).
 
+Community contributions, including bug reports, feature requests, and code contributions, are welcomed at [https://github.com/lanl/scico](https://github.com/lanl/scico).
+
 ![Solving a multi-channel tomography problem with SCICO.\label{fig:flow_chart}](figures/flow_chart.pdf){ width=100% }
+
 
 # Statement of Need
 
@@ -55,296 +60,29 @@ In traditional imaging, the burden of image formation is placed on physical comp
 The goal of SCICO is to provide a general research tool for computational imaging, with a particular focus on scientific imaging applications, which are particularly underrepresented in the existing range of open-source packages in this area. While a number of other packages overlap somewhat in functionality with SCICO, only a few support execution of the same code on both CPU and GPU devices, and we are not aware of any that support just-in-time compilation and automatic gradient computation, which is invaluable in computational imaging. SCICO provides all three of these valuable features by being built on top of [JAX](https://jax.readthedocs.io/en/latest/) rather than [NumPy](https://numpy.org/).
 
 
-
 # Solving Imaging Inverse Problems in SCICO
 
-This section outlines the steps involved in solving an inverse problem,
-and shows how each concept maps to a component of SCICO.
-
-
-## Forward Modeling
-
-In order to solve a computational imaging problem we need to know how the image we wish to reconstruct, $\mb{x}$, is related to the data that we can measure, $\mb{y}$.
[…]
