+++ docs/joss-paper/long-paper.md
@@ -1,0 +1,350 @@
+---
+title: 'Scientific Computational Imaging Code (SCICO)'
+tags:
+  - Python
+  - computational imaging
+  - scientific imaging
+  - inverse problems
+header-includes:
+  - \usepackage{amsmath}
+  - \newcommand{\mb}[1]{\mathbf{#1}}
+  - \DeclareMathOperator*{\argmin}{arg\,min}
+authors:
+  - name: Thilo Balke
+    orcid: 0000-0002-1716-5923
+    affiliation: "1, 2"
+  - name: Fernando Davis
+    affiliation: "1, 3"
+  - name: Cristina Garcia-Cardona
+    orcid: 0000-0002-5641-3491
+    affiliation: 1
+  - name: Soumendu Majee
+    orcid: 0000-0002-8442-2897
+    affiliation: 4
+  - name: Michael McCann
+    orcid: 0000-0001-7645-252X
+    affiliation: 1
+  - name: Brendt Wohlberg
+    orcid: 0000-0002-4767-1843
+    affiliation: 1
+affiliations:
+ - name: Los Alamos National Laboratory
+   index: 1
+ - name: Purdue University
+   index: 2
+ - name: University of Puerto Rico-Mayaguez
+   index: 3
+ - name: Samsung Research America
+   index: 4
+date: 20 August 2022
+bibliography: paper.bib
+---
+
+# Summary
+
+Scientific Computational Imaging Code (SCICO) is a Python package for solving the inverse problems that arise in scientific imaging applications. Its primary focus is providing methods for solving ill-posed inverse problems by using an appropriate prior model of the reconstruction space. SCICO includes a growing suite of operators, cost functionals, regularizers, and optimization routines that may be combined to solve a wide range of problems, and is designed so that it is easy to add new building blocks. SCICO is built on top of [JAX](https://jax.readthedocs.io/en/latest/) rather than [NumPy](https://numpy.org/), enabling GPU/TPU acceleration, just-in-time compilation, and automatic gradient functionality,
+which is used to automatically compute the adjoints of linear operators.
+An example of how to solve a multi-channel tomography problem with SCICO is shown in \autoref{fig:flow_chart}. The SCICO source code is available from [GitHub](https://github.com/lanl/scico), and pre-built packages are available from [PyPI](https://github.com/lanl/scico). It has extensive [online documentation](https://scico.rtfd.io/), including API documentation and usage examples, which can be run online at [Google Colab](https://colab.research.google.com/github/lanl/scico-data/blob/colab/notebooks/index.ipynb) and [binder](https://mybinder.org/v2/gh/lanl/scico-data/binder?labpath=notebooks%2Findex.ipynb).
+
+![Solving a multi-channel tomography problem with SCICO.\label{fig:flow_chart}](figures/flow_chart.pdf){ width=100% }
+
+# Statement of Need
+
+In traditional imaging, the burden of image formation is placed on physical components, such as a lens, with the resulting image being taken from the sensor with minimal processing. In computational imaging, in contrast, the burden of image formation is shared with or shifted to computation, with the resulting image typically being very different from the measured data. Common examples of computational imaging include demosaicing in consumer cameras, computed tomography and magnetic resonance imaging in medicine, and synthetic aperture radar in remote sensing. This is an active and growing area of research, and many of these problems have common properties that could be supported by shared implementations of solution components.
+
[…]
