### 2022-08-24 #4690 @bwohlberg
JORS was one of the more promising ones I looked at, but my impression from their template for software description papers is that they also expect something much shorter than our paper. But if you think I'm being overly pessimistic, perhaps it's worth asking them for clarification.

As a matter of interest, why the strict length constraint for JOSS? I'm sure many other authors would also be interested in a venue for more SciPy-like papers.
