### 2022-08-25 #4690 @bwohlberg
Yes, please. We may consider publishing a much longer paper in the future, but that would be a completely independent document.
