### 2022-08-25 9d2a01e
Improve docs and type annotations
