+++ scico/operator/_operator.py
@@ -26,23 +26,23 @@ from scico.numpy.util import is_nested, shape_to_size
 from scico.typing import BlockShape, DType, JaxArray, Shape
 
 
-def _wrap_mul_div_scalar(func):
-    r"""Wrapper function for defining mul, rmul, and truediv.
+def _wrap_mul_div_scalar(func: Callable) -> Callable:
+    r"""Wrapper function for multiplication and division operators.
 
-    Wrapper function for defining mul, rmul, and truediv between a scalar
-    and an Operator.
+    Wrapper function for defining `__mul__`, `__rmul__`, and
+    `__truediv__` between a scalar and an `Operator`.
 
     If one of these binary operations are called in the form
-    binop(Operator, other) and 'b' is a scalar, specialized
-    Operator constructors can be called.
+    `binop(Operator, other)` and 'b' is a scalar, specialized
+    :class:`.Operator` constructors can be called.
 
     Args:
-        func: should be either .__mul__(), .__rmul__(),
-           or .__truediv__().
+        func: should be either `.__mul__()`, `.__rmul__()`,
+           or `.__truediv__()`.
 
     Raises:
-        TypeError: A binop with the form binop(Operator, other) is
-        called and other is not a scalar.
+        TypeError: If a binop with the form `binop(Operator, other)` is
+        called and `other` is not a scalar.
     """
 
     @wraps(func)
@@ -85,11 +85,11 @@ output_dtype : {self.output_dtype}
                 Defaults to ``None``. If ``None``, `output_shape` is
                 determined by evaluating `self.__call__` on an input
                 array of zeros.
-            eval_fn: Function used in evaluating this Operator.
+            eval_fn: Function used in evaluating this :class:`.Operator`.
                 Defaults to ``None``. If ``None``, then `self.__call__`
                 must be defined in any derived classes.
             input_dtype: `dtype` for input argument.
-                Defaults to ``float32``. If `Operator` implements
+                Defaults to ``float32``. If :class:`.Operator` implements
                 complex-valued operations, this must be ``complex64`` for
                 proper adjoint and gradient calculation.
             output_dtype: `dtype` for output argument.
@@ -97,9 +97,9 @@ output_dtype : {self.output_dtype}
                 determined by evaluating `self.__call__` on an input
                 array of zeros.
             jit: If ``True``, call :meth:`Operator.jit()` on this
-                Operator to jit the forward, adjoint, and gram functions.
-                Same as calling :meth:`Operator.jit` after the Operator
-                is created.
+                :class:`.Operator` to jit the forward, adjoint, and gram
+                functions. Same as calling :meth:`Operator.jit` after the
+                :class:`.Operator` is created.
         """
 
         #: Shape of input array or :class:`.BlockArray`.
@@ -172,11 +172,20 @@ output_dtype : {self.output_dtype}
         r"""Evaluate this Operator at the point :math:`\mb{x}`.
 
         Args:
-            x: Point at which to evaluate this Operator. If `x` is a
-               :class:`DeviceArray` or :class:`.BlockArray`, must have
-               `shape == self.input_shape`. If `x` is a
-               :class:`.Operator` or :class:`.LinearOperator`, must have
-               `x.output_shape == self.input_shape`.
+            x: Point at which to evaluate this :class:`.Operator`. If `x`
+               is a :class:`DeviceArray` or :class:`.BlockArray`, it must
+               have `shape == self.input_shape`. If `x` is a
+               :class:`.Operator` or :class:`.LinearOperator`, it must
+               have `x.output_shape == self.input_shape`.
+
+        Returns:
+             :class:`.Operator` evaluated at `x`.
+
+        Raises:
+            ValueError: If the `input_shape` attribute of the
+                :class:`.Operator` is not equal to the input array shape,
[…]
