+++ scico/linop/_linop.py
@@ -30,28 +30,29 @@ from scico.typing import BlockShape, DType, JaxArray, Shape
 
 
 def _wrap_add_sub(func: Callable, op: Callable) -> Callable:
-    r"""Wrapper function for defining __add__, __sub__.
+    r"""Wrapper function for defining `__add__`, `__sub__`.
 
-    Wrapper function for defining __add__, __sub__ between LinearOperator
-    and other objects.
+    Wrapper function for defining `__add__`,` __sub__` between
+    :class:`LinearOperator` and other objects.
 
     Handles shape checking and dispatching based on operand types:
-    - If one of the two operands is an Operator, an Operator is returned.
-    - If both operands are LinearOperators of different types, a generic
-      LinearOperator is returned.
-    - If both operands are LinearOperators of the same type, a special
-      constructor can be called
+
+    - If one of the two operands is an :class:`.Operator`, an
+      :class:`.Operator` is returned.
+    - If both operands are :class:`LinearOperator` of different types,
+      a generic :class:`LinearOperator` is returned.
+    - If both operands are :class:`LinearOperator` of the same type, a
+      special constructor can be called
 
     Args:
-        func: should be either .__add__() or .__sub__().
+        func: should be either `.__add__` or `.__sub__`.
         op: functional equivalent of func, ex. op.add for func =
-           __add__.
+           `__add__`.
 
     Raises:
-        ValueError: The shape of both operators does not match.
-        TypeError: One of the two operands is not an Operator
-        or LinearOperator.
-
+        ValueError: If the shape of both operators does not match.
+        TypeError: If one of the two operands is not an
+            :class:`.Operator` or :class:`LinearOperator`.
     """
 
     @wraps(func)
@@ -90,7 +91,7 @@ def _wrap_add_sub(func: Callable, op: Callable) -> Callable:
 
 
 class LinearOperator(Operator):
-    """Generic Linear Operator base class"""
+    """Generic linear operator base class"""
 
     def __init__(
         self,
@@ -109,25 +110,27 @@ class LinearOperator(Operator):
                 Defaults to ``None``. If ``None``, `output_shape` is
                 determined by evaluating `self.__call__` on an input
                 array of zeros.
-            eval_fn: Function used in evaluating this LinearOperator.
-                Defaults to ``None``. If ``None``, then `self.__call__`
-                must be defined in any derived classes.
+            eval_fn: Function used in evaluating this
+                :class:`LinearOperator`. Defaults to ``None``. If
+                ``None``, then `self.__call__` must be defined in any
+                derived classes.
             adj_fn: Function used to evaluate the adjoint of this
-                LinearOperator. Defaults to ``None``. If ``None``, the
-                adjoint is not set, and the :meth:`._set_adjoint`
-                will be called silently at the first :meth:`.adj` call or
-                can be called manually.
+                :class:`LinearOperator`. Defaults to ``None``. If
+                ``None``, the adjoint is not set, and the
+                :meth:`._set_adjoint` will be called silently at the
+                first :meth:`.adj` call or can be called manually.
             input_dtype: `dtype` for input argument.
-                Defaults to ``float32``. If `LinearOperator` implements
-                complex-valued operations, this must be ``complex64`` for
-                proper adjoint and gradient calculation.
+                Defaults to ``float32``. If :class:`LinearOperator`
[…]
