### 2022-08-25 0a54d1e
Minor test cleanup
