### 2022-08-26 b1da747
Move Slice linop to _func module
