+++ scico/linop/_func.py
@@ -9,11 +9,12 @@
 from typing import Any, Callable, Optional, Union
 
 import scico.numpy as snp
-from scico.typing import BlockShape, DType, Shape
+from scico.numpy.util import indexed_shape, is_nested
+from scico.typing import ArrayIndex, BlockShape, DType, JaxArray, Shape
 
 from ._linop import LinearOperator
 
-__all__ = ["operator_from_function", "Tranpose", "Sum", "Pad", "Reshape"]
+__all__ = ["operator_from_function", "Tranpose", "Sum", "Pad", "Reshape", "Slice"]
 
 
 def linop_from_function(f: Callable, classname: str, f_name: Optional[str] = None):
@@ -78,3 +79,52 @@ Transpose = linop_from_function(snp.transpose, "Transpose", "scico.numpy.transpo
 Reshape = linop_from_function(snp.reshape, "Reshape")
 Pad = linop_from_function(snp.pad, "Pad", "scico.numpy.pad")
 Sum = linop_from_function(snp.sum, "Sum")
+
+
+class Slice(LinearOperator):
+    """A linear operator for slicing an array."""
+
+    def __init__(
+        self,
+        idx: ArrayIndex,
+        input_shape: Union[Shape, BlockShape],
+        input_dtype: DType = snp.float32,
+        jit: bool = True,
+        **kwargs,
+    ):
+        r"""
+        This operator may be applied to either a :any:`JaxArray` or a
+        :class:`.BlockArray`. In the latter case, parameter `idx` must
+        conform to the
+        :ref:`BlockArray indexing requirements <blockarray_indexing>`.
+
+        Args:
+            idx: An array indexing expression, as generated by
+                :data:`numpy.s_`, for example.
+            input_shape: Shape of input :any:`JaxArray` or :class:`.BlockArray`.
+            input_dtype: `dtype` for input argument.
+                Defaults to ``float32``. If this LinearOperator implements
+                complex-valued operations, this must be ``complex64`` for
+                proper adjoint and gradient calculation.
+            jit: If ``True``, jit the evaluation, adjoint, and gram
+               functions of the LinearOperator.
+        """
+
+        output_shape: Union[Shape, BlockShape]
+        if is_nested(input_shape):
+            output_shape = input_shape[idx]  # type: ignore
+        else:
+            output_shape = indexed_shape(input_shape, idx)
+
+        self.idx: ArrayIndex = idx
+        super().__init__(
+            input_shape=input_shape,
+            output_shape=output_shape,
+            input_dtype=input_dtype,
+            output_dtype=input_dtype,
+            jit=jit,
+            **kwargs,
+        )
+
+    def _eval(self, x: JaxArray) -> JaxArray:
+        return x[self.idx]
