+++ scico/linop/__init__.py
@@ -13,10 +13,10 @@ from ._circconv import CircularConvolve
 from ._convolve import Convolve, ConvolveByX
 from ._dft import DFT
 from ._diff import FiniteDifference, SingleAxisFiniteDifference
-from ._func import Pad, Reshape, Sum, Transpose, linop_from_function
+from ._func import Pad, Reshape, Slice, Sum, Transpose, linop_from_function
 from ._linop import ComposedLinearOperator, LinearOperator
 from ._matrix import MatrixOperator
-from ._misc import Diagonal, Identity, Slice
+from ._misc import Diagonal, Identity
 from ._stack import DiagonalStack, VerticalStack
 from ._util import operator_norm, power_iteration, valid_adjoint
 
