### 2022-08-26 842f3ee
Move slice linop tests
