### 2022-08-26 8e8769e
Add Crop linop. Resolves #231
