+++ scico/linop/_func.py
@@ -6,15 +6,16 @@
 
 """Linear operators constructed from functions."""
 
-from typing import Any, Callable, Optional, Union
+from typing import Any, Callable, Optional, Sequence, Union
 
 import scico.numpy as snp
+from scico._autograd import linear_adjoint
 from scico.numpy.util import indexed_shape, is_nested
 from scico.typing import ArrayIndex, BlockShape, DType, JaxArray, Shape
 
 from ._linop import LinearOperator
 
-__all__ = ["operator_from_function", "Tranpose", "Sum", "Pad", "Reshape", "Slice"]
+__all__ = ["operator_from_function", "Tranpose", "Sum", "Crop", "Pad", "Reshape", "Slice"]
 
 
 def linop_from_function(f: Callable, classname: str, f_name: Optional[str] = None):
@@ -79,6 +80,52 @@ Transpose = linop_from_function(snp.transpose, "Transpose", "scico.numpy.transpo
 Reshape = linop_from_function(snp.reshape, "Reshape")
 Pad = linop_from_function(snp.pad, "Pad", "scico.numpy.pad")
 Sum = linop_from_function(snp.sum, "Sum")
+
+
+class Crop(LinearOperator):
+    """A linear operator for cropping an array."""
+
+    def __init__(
+        self,
+        crop_width: Union[int, Sequence],
+        input_shape: Shape,
+        input_dtype: DType = snp.float32,
+        jit: bool = True,
+        **kwargs,
+    ):
+        r"""
+        Args:
+            crop_width: Specify the crop width using the same format as
+                the `pad_width` parameter of :func:`snp.pad`.
+            input_shape: Shape of input :any:`JaxArray.
+            input_dtype: `dtype` for input argument.
+                Defaults to ``float32``. If this LinearOperator implements
+                complex-valued operations, this must be ``complex64`` for
+                proper adjoint and gradient calculation.
+            jit: If ``True``, jit the evaluation, adjoint, and gram
+               functions of the LinearOperator.
+        """
+
+        self.crop_width = crop_width
+        # The crop function is defined as the adjoint of snp.pad
+        pad = lambda x: snp.pad(x, pad_width=crop_width)
+        # The output shape of this operator is the input shape of the corresponding
+        # pad operation of which it is the adjoint. Since we don't know this output
+        # shape, we assume that it can be computed by subtracting the difference in
+        # output and input shapes resulting from applying the pad operator to the
+        # input shape of this operator.
+        tmp = pad(snp.zeros(input_shape, dtype=input_dtype))
+        output_shape = tuple(2 * snp.array(input_shape) - snp.array(tmp.shape))
+        pad_adjoint = linear_adjoint(pad, snp.zeros(output_shape, dtype=input_dtype))
+        super().__init__(
+            input_shape=input_shape,
+            input_dtype=input_dtype,
+            eval_fn=lambda x: pad_adjoint(x)[0],
+            output_shape=output_shape,
+            output_dtype=input_dtype,
+            jit=jit,
+            **kwargs,
+        )
 
 
 class Slice(LinearOperator):
