+++ scico/linop/__init__.py
@@ -14,7 +14,7 @@ from ._convolve import Convolve, ConvolveByX
 from ._dft import DFT
 from ._diag import Diagonal, Identity
 from ._diff import FiniteDifference, SingleAxisFiniteDifference
-from ._func import Pad, Reshape, Slice, Sum, Transpose, linop_from_function
+from ._func import Crop, Pad, Reshape, Slice, Sum, Transpose, linop_from_function
 from ._linop import ComposedLinearOperator, LinearOperator
 from ._matrix import MatrixOperator
 from ._stack import DiagonalStack, VerticalStack
