### 2022-08-26 7a29f41
Fix docstring typo
