+++ scico/linop/_stack.py
@@ -50,7 +50,7 @@ def is_collapsible(shapes: Sequence[Union[Shape, BlockShape]]) -> bool:
 
 def is_blockable(shapes: Sequence[Union[Shape, BlockShape]]) -> TypeGuard[Union[Shape, BlockShape]]:
     """Return ``True`` if the list of shapes represent arrays that
-    can be combined into a BlockArray, i.e., none are nested."""
+    can be combined into a :class:`BlockArray`, i.e., none are nested."""
     return not any(is_nested(s) for s in shapes)
 
 
@@ -67,10 +67,11 @@ class VerticalStack(LinearOperator):
         r"""
         Args:
             ops: Operators to stack.
-            collapse: If ``True`` and the output would be a `BlockArray`
-                with shape ((m, n, ...), (m, n, ...), ...), the output is
-                instead a `DeviceArray` with shape (S, m, n, ...) where S
-                is the length of `ops`. Defaults to ``True``.
+            collapse: If ``True`` and the output would be a
+                :class:`BlockArray` with shape ((m, n, ...), (m, n, ...),
+                ...), the output is instead a `DeviceArray` with shape
+                (S, m, n, ...) where S is the length of `ops`. Defaults
+                to ``True``.
             jit: see `jit` in :class:`LinearOperator`.
 
         """
@@ -106,14 +107,14 @@ class VerticalStack(LinearOperator):
         input_shapes = [op.shape[1] for op in ops]
         if not all(input_shapes[0] == s for s in input_shapes):
             raise ValueError(
-                "Expected all `LinearOperator`s to have the same input shapes, "
+                "Expected all LinearOperators to have the same input shapes, "
                 f"but got {input_shapes}"
             )
 
         input_dtypes = [op.input_dtype for op in ops]
         if not all(input_dtypes[0] == s for s in input_dtypes):
             raise ValueError(
-                "Expected all `LinearOperator`s to have the same input dtype, "
+                "Expected all LinearOperators to have the same input dtype, "
                 f"but got {input_dtypes}."
             )
 
@@ -122,7 +123,7 @@ class VerticalStack(LinearOperator):
 
         output_dtypes = [op.output_dtype for op in ops]
         if not np.all(output_dtypes[0] == s for s in output_dtypes):
-            raise ValueError("Expected all `LinearOperator`s to have the same output dtype.")
+            raise ValueError("Expected all LinearOperators to have the same output dtype.")
 
     def _eval(self, x: JaxArray) -> Union[JaxArray, BlockArray]:
         if self.collapsible and self.collapse:
