+++ docs/source/references.bib
@@ -96,6 +96,17 @@
   year =	 {2018}
 }
 
+@InProceedings {benning-2016-preconditioned,
+  title =	 {Preconditioned {ADMM} with nonlinear operator
+                  constraint},
+  author =	 {Benning, Martin and Knoll, Florian and
+                  Sch{\"o}nlieb, Carola-Bibiane and Valkonen, Tuomo},
+  booktitle =	 {IFIP Conference on System Modeling and Optimization (CSMO) 2015},
+  pages =	 {117--126},
+  year =	 2016,
+  doi =		 {10.1007/978-3-319-55795-3_10}
+}
+
 @Article {boyd-2010-distributed,
   title =	 {Distributed optimization and statistical learning
                   via the alternating direction method of multipliers},
