### 2022-08-30 cd7bd14
Update comments in Flax train
