### 2022-08-30 41e1f27
Refactor Flax train functionality
