+++ scico/flax/train/train.py
@@ -63,10 +63,20 @@ class ConfigDict(TypedDict):
     base_learning_rate: float
     lr_decay_rate: float
     warmup_epochs: int
-    num_train_steps: int
     steps_per_eval: int
     log_every_steps: int
     steps_per_epoch: int
+    steps_per_checkpoint: int
+    log: bool
+    workdir: str
+    checkpointing: bool
+    return_state: bool
+    lr_schedule: Callable
+    criterion: Callable
+    create_train_state: Callable
+    train_step_fn: Callable
+    eval_step_fn: Callable
+    post_lst: List[Callable]
 
 
 class ModelVarDict(TypedDict):
@@ -333,7 +343,7 @@ def _train_step(
 ) -> Tuple[TrainState, MetricsDict]:
     """Perform a single data parallel training step. Assumes sharded batched data.
 
-    This function is intended to be used via :meth:`train_and_evaluate`, not directly.
+    This function is intended to be used via :class:`BasicFlaxTrainer`, not directly.
 
     Args:
         state: Flax train state which includes the
@@ -438,7 +448,7 @@ def _train_step_post(
     condition, etc.) is applied after the gradient update.
     Assumes sharded batched data.
 
-    This function is intended to be used via :meth:`train_and_evaluate`, not directly.
+    This function is intended to be used via :class:`BasicFlaxTrainer`, not directly.
 
     Args:
         state: Flax train state which includes the
@@ -471,7 +481,7 @@ def _eval_step(state: TrainState, batch: DataSetDict) -> MetricsDict:
     """Evaluate current model state. Assumes sharded
     batched data.
 
-    This function is intended to be used via :meth:`train_and_evaluate` or :meth:`only_evaluate`, not directly.
+    This function is intended to be used via :class:`BasicFlaxTrainer` or :meth:`only_evaluate`, not directly.
 
     Args:
         state: Flax train state which includes the
@@ -557,59 +567,257 @@ def stats_obj() -> Tuple[IterationStats, Callable]:
     return itstat_object, itstat_insert_func
 
 
-def train_and_evaluate(
-    config: ConfigDict,
-    workdir: str,
-    model: ModuleDef,
-    train_ds: DataSetDict,
-    test_ds: DataSetDict,
-    create_lr_schedule: Callable = create_cnst_lr_schedule,
-    create_train_state: Callable = create_basic_train_state,
-    criterion: Callable = mse_loss,
-    train_step_fn: Callable = _train_step,
-    eval_step_fn: Callable = _eval_step,
-    post_lst: Optional[List[Callable]] = None,
-    variables0: Optional[ModelVarDict] = None,
-    checkpointing: bool = False,
-    log: bool = False,
-    return_state: bool = False,
-) -> Tuple[Union[ModelVarDict, TrainState], IterationStats]:
-    """Execute model training and evaluation loop.
+class BasicFlaxTrainer:
+    """Class for encapsulating Flax training configuration and execution."""
 
-    Args:
-        config: Hyperparameter configuration.
-        workdir: Directory to write checkpoints.
-        model: Flax model to train.
-        train_ds: Dictionary of training data (includes images
-            and labels).
-        test_ds: Dictionary of testing data (includes images
-            and labels).
-        create_lr_schedule: A function that creates an Optax
-            learning rate schedule. Default:
-            :meth:`create_cnst_lr_schedule`.
-        create_train_state: A function that creates a Flax train state and initializes it. A train state object helps to keep optimizer and module functionality grouped for training. Default:
-            :meth:`create_basic_train_state`.
-        criterion: A function that specifies the loss being minimized in training. Default: :meth:`mse_loss`.
[…]
