### 2022-08-30 2f0c408
Improve docstrings in train functionality
