### 2022-08-30 fe2e3a6
Update spectral utils in Flax train functionality
