### 2022-08-30 ed5b2e6
Update Flax __init__
