### 2022-08-30 fbf2fb8
Update and add more Flax train tests
