+++ scico/test/flax/test_train.py
@@ -1,20 +1,36 @@
+import functools
 import os
 
 import numpy as np
 
 import jax
 
+import optax
 import pytest
 
+from flax import jax_utils
 from scico import flax as sflax
 from scico import random
+from scico.flax.train.clu_utils import flatten_dict
 from scico.flax.train.input_pipeline import IterateData, prepare_data
 from scico.flax.train.train import (
+    ArgumentStruct,
+    _apply_fn,
+    _eval_step,
+    _train_step,
+    _train_step_post,
+    clip_positive,
+    clip_range,
     compute_metrics,
+    construct_traversal,
+    create_basic_train_state,
     create_cnst_lr_schedule,
     create_cosine_lr_schedule,
     create_exp_lr_schedule,
+    initialize,
     mse_loss,
+    stats_obj,
+    sync_batch_stats,
 )
 
 os.environ["XLA_FLAGS"] = "--xla_force_host_platform_device_count=8"
@@ -54,25 +70,22 @@ class SetupTest:
         self.test_ds = {"image": xt, "label": xt}
 
         # Model configuration
-        self.mconf = {
+        self.model_conf = {
             "depth": 2,
             "num_filters": 16,
             "block_depth": 2,
         }
 
         # Training configuration
-        self.dconf: sflax.ConfigDict = {
+        self.train_conf: sflax.ConfigDict = {
             "seed": 0,
             "opt_type": "ADAM",
             "momentum": 0.9,
-            "batch_size": 16,
-            "num_epochs": 2,
+            "batch_size": self.bsize,
+            "num_epochs": 1,
             "base_learning_rate": 1e-3,
             "lr_decay_rate": 0.95,
             "warmup_epochs": 0,
-            "num_train_steps": -1,
-            "steps_per_eval": -1,
-            "steps_per_epoch": 1,
             "log_every_steps": 1000,
         }
 
@@ -154,111 +167,466 @@ def test_compute_metrics(testobj):
     assert mtrcs["snr"] < 5e-4
 
 
-def test_count_parameters(testobj):
-    model = sflax.ResNet(testobj.mconf["depth"], testobj.chn, testobj.mconf["num_filters"])
-
-    key = jax.random.PRNGKey(seed=1234)
-    input_shape = (1, testobj.N, testobj.N, testobj.chn)
-    variables = model.init({"params": key}, np.ones(input_shape, model.dtype))
-
-    filter_sz = model.kernel_size[0] * model.kernel_size[1]
-    # filter parameters output layer
-    sum_manual_params = filter_sz * testobj.mconf["num_filters"] * testobj.chn
-    # bias and scale of batch normalization output layer
-    sum_manual_params += testobj.chn * 2
-    # mean and bar of batch normalization output layer
-    sum_manual_bst = testobj.chn * 2
-    chn_prev = 1
-    for i in range(testobj.mconf["depth"] - 1):
-        # filter parameters
-        sum_manual_params += filter_sz * testobj.mconf["num_filters"] * chn_prev
-        # bias and scale of batch normalization
-        sum_manual_params += testobj.mconf["num_filters"] * 2
-        # mean and bar of batch normalization
-        sum_manual_bst += testobj.mconf["num_filters"] * 2
-        chn_prev = testobj.mconf["num_filters"]
-
-    total_nvar_params = sflax.count_parameters(variables["params"])
-    total_nvar_bst = sflax.count_parameters(variables["batch_stats"])
-
-    assert total_nvar_params == sum_manual_params
-    assert total_nvar_bst == sum_manual_bst
-
-
[…]
