### 2022-08-30 8e0f820
Update and add more spectral utils tests
