+++ scico/test/flax/test_utils.py
@@ -1,4 +1,5 @@
 from functools import partial
+from typing import Any, Tuple
 
 import numpy as np
 
@@ -6,6 +7,8 @@ import jax
 
 import pytest
 
+from flax.linen import Conv
+from flax.linen.module import Module, compact
 from scico import flax as sflax
 from scico import linop
 from scico.flax.train.train import construct_traversal
@@ -30,19 +33,65 @@ def test_l2_normalize():
     np.testing.assert_allclose(l2n_jnp, l2n_util, rtol=eps)
 
 
-@pytest.mark.parametrize("kernel_shape", [(3, 3, 1, 1), (11, 11, 1, 1)])
-def test_conv(kernel_shape):
+@pytest.mark.parametrize("kernel_size", [(3, 3, 1, 1), (11, 11, 1, 1)])
+def test_conv(kernel_size):
     key = jax.random.PRNGKey(97531)
-    kernel, key = randn(kernel_shape, dtype=np.float32, key=key)
+    kernel, key = randn(kernel_size, dtype=np.float32, key=key)
 
-    input_shape = (1, 128, 128, 1)
-    x, key = randn(input_shape, dtype=np.float32, key=key)
+    input_size = (1, 128, 128, 1)
+    x, key = randn(input_size, dtype=np.float32, key=key)
 
-    y = jax.scipy.signal.convolve(x.squeeze(), jax.numpy.flip(kernel).squeeze(), mode="same")
+    pads = (
+        [(0, 0)]
+        + [(kernel_size[0] // 2, kernel_size[0] // 2)]
+        + [(kernel_size[1] // 2, kernel_size[1] // 2)]
+        + [(0, 0)]
+    )
+    xext = np.pad(x, pads, mode="wrap")
+
+    y = jax.scipy.signal.convolve(xext.squeeze(), jax.numpy.flip(kernel).squeeze(), mode="valid")
 
     y_util = conv(x, kernel).squeeze()
 
     np.testing.assert_allclose(y, y_util)
+
+
+class CNN(Module):
+    kernel_size: Tuple[int, int]
+    kernel0: Any
+
+    @compact
+    def __call__(self, x):
+        def kinit_wrap(rng, shape, dtype=np.float32):
+            return np.array(self.kernel0, dtype)
+
+        return Conv(
+            features=1,
+            kernel_size=self.kernel_size,
+            use_bias=False,
+            padding="CIRCULAR",
+            kernel_init=kinit_wrap,
+        )(x)
+
+
+@pytest.mark.parametrize("kernel_size", [(3, 3, 1, 1), (11, 11, 1, 1)])
+def test_conv_layer(kernel_size):
+    key = jax.random.PRNGKey(12345)
+    kernel, key = randn(kernel_size, dtype=np.float32, key=key)
+
+    input_size = (1, 128, 128, 1)
+    x, key = randn(input_size, dtype=np.float32, key=key)
+
+    rng = jax.random.PRNGKey(42)
+    model = CNN(kernel_size=kernel_size[:2], kernel0=kernel)
+    variables = model.init(rng, np.zeros(x.shape))
+    prms = variables["params"]
+    np.testing.assert_allclose(prms["Conv_0"]["kernel"], kernel)
+
+    y_layer = model.apply(variables, x)
+    y_util = conv(x, kernel)
+
+    np.testing.assert_allclose(y_layer, y_util)
 
 
 @pytest.mark.parametrize("input_shape", [(8,), (128,)])
@@ -56,7 +105,7 @@ def test_spectral_norm(input_shape):
     x, key = randn(input_shape, dtype=np.float32, key=key)
     mu_util = estimate_spectral_norm(lambda x: D @ x, x.shape, n_steps=200)
 
[…]
