### 2022-08-30 089819e
Update Flax examples utils and tests
