+++ scico/flax/examples.py
@@ -205,7 +205,7 @@ def ray_distributed_data_generation(
     if not have_ray:
         raise RuntimeError("Package ray is required for use of this function.")
 
-    ray.init()
+    ray.init(local_mode=True, ignore_reinit_error=True)
 
     @ray.remote
     def data_gen(seed, size, ndata, imgf):
@@ -214,6 +214,8 @@ def ray_distributed_data_generation(
     ar = ray.available_resources()
     # Usage of half available CPU resources.
     nproc = max(int(ar["CPU"]) // 2, 1)
+    if nproc > nimg:
+        nproc = nimg
     if nproc > 1 and nimg % nproc > 0:
         raise ValueError(
             f"Number of images to generate ({nimg}) must be divisible by the number of available devices ({nproc})"
@@ -237,7 +239,7 @@ def ct_data_generation(
     imgfunc: Callable = generate_foam2_images,
     seed: int = 1234,
     verbose: bool = False,
-) -> Tuple[Array, ...]:  # pragma: no cover
+) -> Tuple[Array, ...]:
     """
     Generate CT data.
 
@@ -473,15 +475,15 @@ def load_ct_data(
     return trdt, ttdt
 
 
-def foam_blur_data_generation(
+def blur_data_generation(
     nimg: int,
     size: int,
     blur_kernel: Array,
     noise_sigma: float,
-    imgfunc: Callable = generate_foam_images,
+    imgfunc: Callable,
     seed: int = 4321,
     verbose: bool = False,
-) -> Tuple[Array, ...]:  # pragma: no cover
+) -> Tuple[Array, ...]:
     """
     Generate blurred data based on xdesign foam structures.
 
@@ -536,7 +538,7 @@ def foam_blur_data_generation(
     # Clip to [0,1] range.
     blurn = jnp.clip(blurn, a_min=0, a_max=1)
 
-    if verbose:
+    if verbose:  # pragma: no cover
         platform = jax.lib.xla_bridge.get_backend().platform
         print(f"{'Platform':28s}{':':4s}{platform}")
         print(f"{'Device count':28s}{':':4s}{jax.device_count()}")
@@ -664,11 +666,12 @@ def load_foam_blur_data(
 
     # Generate new data.
     nimg = train_nimg + test_nimg
-    img, blrn = foam_blur_data_generation(
+    img, blrn = blur_data_generation(
         nimg,
         size,
         blur_kernel,
         noise_sigma,
+        imgfunc=generate_foam_images,
         verbose=verbose,
     )
     # Separate training and testing partitions.
@@ -939,7 +942,7 @@ def reconfigure_images(
         S = np.zeros((num_img, images.shape[1], images.shape[2], C), dtype=dtype)
     else:
         tsfm_crop = CenterCrop(output_size)
-        S = np.zeros((num_img, tsfm.output_size[0], tsfm.output_size[1], C), dtype=dtype)
+        S = np.zeros((num_img, tsfm_crop.output_size[0], tsfm_crop.output_size[1], C), dtype=dtype)
 
     # Convert to gray scale and/or crop.
     for i in range(S.shape[0]):
@@ -1069,7 +1072,7 @@ def build_image_dataset(
     return train_ds, test_ds
 
 
-def images_read(path: str, ext: str = "jpg") -> Array:
+def images_read(path: str, ext: str = "jpg") -> Array:  # pragma: no cover
     """Read a collection of color images from a set of files in the specified directory.
 
     All files with extension `ext` (i.e. matching glob `*.ext`)
@@ -1160,6 +1163,72 @@ def get_bsds_data(path: str, verbose: bool = False):  # pragma: no cover
     if verbose:
[…]
