### 2022-08-30 52bb330
Update Flax example scripts
