+++ examples/scripts/denoise_datagen_bsds.py
@@ -77,7 +77,6 @@ fig.colorbar(
     ax=axes,
     shrink=0.5,
     pad=0.05,
-    label="Arbitrary Units",
 )
 fig.show()
 
