+++ examples/scripts/ct_astra_unet_train_foam2.py
@@ -63,26 +63,23 @@ at each pooling (unpooling) operation.
 Better performance may be obtained by increasing depth, block depth, number of filters or training epochs,
 but may require longer training times.
 """
-batch_size = 16
-epochs = 200
 # model configuration
-mconf = {
+model_conf = {
     "depth": 2,
     "num_filters": 64,
     "block_depth": 2,
 }
 # training configuration
-dconf: sflax.ConfigDict = {
+train_conf: sflax.ConfigDict = {
     "seed": 0,
     "opt_type": "SGD",
     "momentum": 0.9,
-    "batch_size": batch_size,
-    "num_epochs": epochs,
+    "batch_size": 16,
+    "num_epochs": 200,
     "base_learning_rate": 1e-2,
     "warmup_epochs": 0,
-    "num_train_steps": -1,
-    "steps_per_eval": -1,
     "log_every_steps": 1000,
+    "log": True,
 }
 
 """
@@ -90,24 +87,33 @@ Construct UNet model.
 """
 channels = train_ds["image"].shape[-1]
 model = sflax.UNet(
-    depth=mconf["depth"],
+    depth=model_conf["depth"],
     channels=channels,
-    num_filters=mconf["num_filters"],
-    block_depth=mconf["block_depth"],
+    num_filters=model_conf["num_filters"],
+    block_depth=model_conf["block_depth"],
 )
 
 """
 Run training loop.
 """
 workdir = os.path.join(os.path.expanduser("~"), ".cache", "scico", "examples", "unet_ct_out")
+train_conf["workdir"] = workdir
 print(f"{'JAX process: '}{jax.process_index()}{' / '}{jax.process_count()}")
 print(f"{'JAX local devices: '}{jax.local_devices()}")
 
-start_time = time()
-modvar, stats_object = sflax.train_and_evaluate(
-    dconf, workdir, model, train_ds, test_ds, checkpointing=False, log=True
+
+# Construct training object
+trainer = sflax.BasicFlaxTrainer(
+    train_conf,
+    model,
+    train_ds,
+    test_ds,
 )
+
+start_time = time()
+modvar, stats_object = trainer.train()
 time_train = time() - start_time
+
 
 """
 Evaluate on testing data.
@@ -124,7 +130,9 @@ and data fidelity.
 """
 snr_eval = metric.snr(test_ds["label"], output)
 psnr_eval = metric.psnr(test_ds["label"], output)
-print(f"{'UNet training':15s}{'epochs:':2s}{epochs:>5d}{'':21s}{'time[s]:':10s}{time_train:>7.2f}")
+print(
+    f"{'UNet training':15s}{'epochs:':2s}{train_conf['num_epochs']:>5d}{'':21s}{'time[s]:':10s}{time_train:>7.2f}"
+)
 print(
     f"{'UNet testing':15s}{'SNR:':5s}{snr_eval:>5.2f}{' dB'}{'':3s}{'PSNR:':6s}{psnr_eval:>5.2f}{' dB'}{'':3s}{'time[s]:':10s}{time_eval:>7.2f}"
 )

[…]
