### 2022-08-30 fdbba12
Remove parameters in ray initialization in Flax example utils
