### 2022-08-30 8f24924
Change ray initialization for Flax example utils tests
