### 2022-08-30 33989d4
Remove xdesign dependency in Flax ct data generation example util
