### 2022-08-30 3e0a1c5
Update CNN script
