### 2022-08-30 e6e3984
Add install commands
