+++ requirements.txt
@@ -8,5 +8,5 @@ jax>=0.3.0,<=0.3.15
 flax>=0.4.0
 bm3d>=3.0.9
 bm4d
-svmbir>=0.2.9
+svmbir>=0.2.9,<0.3.0
 pyabel>=0.8.5
