### 2022-08-30 31bc5ed
Pin matplotlib to work in Colab
