### 2022-08-30 28c8da3
Change importing of TypedDict in Flax funtionality for compatibility with python 3.7
