### 2022-08-31 2454940
Updated CNN tutorial
