### 2022-08-31 #4690 @vitorsr
> 👋 @vitorsr - Are you able to review this JOSS submission?

I would be happy to.
