### 2022-08-31 #4690 @DanNixon
> wave @DanNixon - Are you able to review this JOSS submission?

I'm happy to.
