+++ requirements.txt
@@ -5,7 +5,7 @@ imageio>=2.17
 matplotlib<=3.1.3
 jaxlib>=0.3.0,<=0.3.15
 jax>=0.3.0,<=0.3.15
-flax>=0.4.0
+flax>=0.4.0,<=0.4.1
 bm3d>=3.0.9
 bm4d
 svmbir>=0.2.9,<0.3.0
