### 2022-08-31 #4690 @danielskatz
👋 @fabian-sp, @glemaitre, @jameschapman19, @Leo-Simpson, @lucaferranti, @robertmartin8, @stsievert, and @Viech - We're looking for one more reviewer for this JOSS submission.  Would one of you be willing to do this?

JOSS publishes articles about open source research software. The submission here is titled: "Scientific Computational Imaging Code (SCICO)" and the submission repository is at: https://github.com/lanl/scico

JOSS is a free, open-source, community driven and developer-friendly online journal (no publisher is seeking to raise revenue from the volunteer labor of researchers!).

The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged.

JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html

Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review).
