### 2022-08-31 #4722 @editorialbot
[REVIEW]: Scientific Computational Imaging Code (SCICO)
**Submitting author:** @bwohlberg (<a href="http://orcid.org/0000-0002-4767-1843">Brendt Wohlberg</a>)
**Repository:** https://github.com/lanl/scico
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.0.2
