+++ scico/flax/train/clu_utils.py
@@ -17,6 +17,10 @@
 #     http://www.apache.org/licenses/LICENSE-2.0
 
 
+import warnings
+
+warnings.simplefilter(action="ignore", category=FutureWarning)
+
 import dataclasses
 from typing import Any, Callable, Dict, List, Mapping, Optional, Sequence, Tuple, Union
 
