### 2022-08-31 d876972
Add extended CNN tutorial
