+++ examples/tutorial/tutorial_CNN_extended.py
@@ -1,0 +1,792 @@
+"""
+# Example: Data-Driven Priors for Inverse Problems
+
+In this example,
+ we will explore how to use CNN models as data-driven priors in a SCICO pipeline for performing computed tomography (CT) reconstruction.
+"""
+
+"""
+## Setup
+To set up your environment, run the cell below.
+
+If you get a popup with 'Warning: This notebook was not authored by Google.', select 'Run anyway'.
+You should see console outputs appearing.
+The install may take several minutes;
+when it is finished, you should see `==done with install==`.
+
+"""
+!pip install -q condacolab
+import condacolab
+condacolab.install()
+
+!pip install git+https://github.com/lanl/scico@cristina/more-flax
+!pip install xdesign
+!conda install -c astra-toolbox astra-toolbox
+
+print('==done with install==')
+
+"""
+## Introduction
+
+Suppose that you are performing CT scans of many similar objects
+ and want to construct a pipeline to rapidly compute the reconstruction
+ of each new measurement.
+For this example, we will use computer-generated foam images
+as the objects we want to image.
+Run the next cell to generate and visualize one of such foams.
+"""
+
+import numpy as np
+
+import matplotlib.pyplot as plt
+from xdesign import Foam, discrete_phantom
+
+from scico import plot
+
+plot.config_notebook_plotting()  # set up plotting
+plt.rcParams["image.cmap"] = "gray"  # set default colormap
+
+np.random.seed(7654)
+
+N = 256  # image size
+x_fm = discrete_phantom(Foam(size_range=[0.075, 0.0025], gap=1e-3, porosity=1), size=N)
+x_gt = x_fm / np.max(x_fm)
+x_gt = np.clip(x_gt, 0, 1.0)
+
+# Plot signal
+fig, ax = plt.subplots()
+ax.imshow(x_gt)
+ax.set_title("Foam")
+fig.show()
+
+"""
+This image shows one foam,
+ but data-driven pipelines assume that you have access to a wealth of relevant data.
+ Let's generate several different foams to use in our CT reconstruction pipeline.
+ Since we are interested in CT reconstruction, we need to generate both images and sinograms.
+
+SCICO provides CT projectors based on Python libraries such as ASTRA and SVMBIR. In this case we will use the
+ASTRA interface (see https://scico.readthedocs.io/en/latest/_autosummary/scico.linop.radon_astra.html).
+
+**Define an ASTRA SCICO CT projector assuming 60 equally spaced projections.**
+"""
+# startq
+
+n_projection = ...  # number of projections
+angles = ...
+A = ...
+# starta
+from scico.linop.radon_astra import TomographicProjector
+
+n_projection = 60  # number of projections
+angles = np.linspace(0, np.pi, n_projection, endpoint=False, dtype=np.float32)  # evenly spaced projection angles
+A = TomographicProjector(x_gt.shape, 1, N, angles)  # Radon transform operator
+# endqa
+
+"""
+Machine learning algorithms are typically very sensitive to the scaling of their inputs.
+For this reason, we normalize the operator `A` by the dimension of the image,
+which, for this operator, makes $ ||Ax|| \approx ||x||$.
+"""
+
+A = A / N
+
+"""
+**Test your operator by computing the sinogram of the generated foam and plotting your results.**
+"""
+# startq
+from mpl_toolkits.axes_grid1 import make_axes_locatable
+
+sino = ...
+
+fig, ax = plot.subplots(nrows=1, ncols=2, figsize=(15, 5))
+plot.imview(..., title="Ground truth", cbar=None, fig=fig, ax=ax[0])
+plot.imview(
+    ...,
+    title="Sinogram",
[…]
