### 2022-08-31 3d2263c
Fix tutorial_CNN_extended
