### 2022-09-01 09eb5ca
Add criterion in metrics inside train step
