### 2022-09-01 66fbf37
Fix errors in extended CNN tutorial script
