### 2022-09-01 ba4ee98
Add metrics option in Flax training
