### 2022-09-01 e1736d1
Update extended CNN tutorial script
