### 2022-09-01 3312ac1
Fix train metrics collection in Flax train
