### 2022-09-02 88026d2
Clean up `Operator` and `LinearOperator` module structure (#329)
* Move definitions of Operator and LinearOperator base classes

* Move definition of linops based on functions to new module

* Move misc linop definitions into new module

* Move utility functions to new module

* Minor docs fix

* Resolve mypy errors

* Trivial edit

* Improve docs and type annotations

* Improve docs

* Resolve #312

* Minor test cleanup

* Move Slice linop to _func module

* Rename _misc.py to _diag.py

* Move slice linop tests

* Add Crop linop. Resolves #231

* Add Crop linop to __all__

* Fix docstring typ
[…]
