+++ scico/linop/__init__.py
@@ -9,27 +9,16 @@
 
 import sys
 
-# isort: off
-from scico._generic_operators import LinearOperator
-from ._linop import (
-    Diagonal,
-    Identity,
-    Pad,
-    Slice,
-    Sum,
-    Transpose,
-    linop_from_function,
-    operator_norm,
-    power_iteration,
-    valid_adjoint,
-)
-from ._matrix import MatrixOperator
-from ._diff import FiniteDifference, SingleAxisFiniteDifference
-from ._convolve import Convolve, ConvolveByX
 from ._circconv import CircularConvolve
+from ._convolve import Convolve, ConvolveByX
 from ._dft import DFT
-from ._stack import VerticalStack, DiagonalStack
-
+from ._diag import Diagonal, Identity
+from ._diff import FiniteDifference, SingleAxisFiniteDifference
+from ._func import Crop, Pad, Reshape, Slice, Sum, Transpose, linop_from_function
+from ._linop import ComposedLinearOperator, LinearOperator
+from ._matrix import MatrixOperator
+from ._stack import DiagonalStack, VerticalStack
+from ._util import operator_norm, power_iteration, valid_adjoint
 
 __all__ = [
     "CircularConvolve",
@@ -43,10 +32,13 @@ __all__ = [
     "DiagonalStack",
     "MatrixOperator",
     "Pad",
+    "Crop",
+    "Reshape",
     "Slice",
     "Sum",
     "Transpose",
     "LinearOperator",
+    "ComposedLinearOperator",
     "linop_from_function",
     "operator_norm",
     "power_iteration",
