+++ docs/source/overview.rst
@@ -25,40 +25,26 @@ Overview
 
 
 
-`SCICO <https://github.com/lanl/scico>`__ is a Python package
-for solving the inverse problems that arise
-in scientific imaging applications.
-Its primary focus is providing methods
-for solving ill-posed inverse problems
-by using an appropriate prior model of the reconstruction space.
-SCICO includes a growing suite of operators, cost
-functionals, regularizers, and optimization routines that may be
-combined to solve a wide range of problems and is designed so that it is
-easy to add new building blocks.
+`SCICO <https://github.com/lanl/scico>`__ is a Python package for solving the inverse problems that arise in scientific imaging applications.  Its primary focus is providing methods for solving ill-posed inverse problems by using an appropriate prior model of the reconstruction space.  SCICO includes a growing suite of operators, cost functionals, regularizers, and optimization routines that may be combined to solve a wide range of problems and is designed so that it is easy to add new building blocks.
 
 SCICO's main advantages are:
 
-   - Operators in SCICO are defined using a simple, NumPy-like syntax,
-     yet automatically benefit from GPU/TPU acceleration and
-     `just-in-time compilation
-     <https://jax.readthedocs.io/en/latest/notebooks/quickstart.html#using-jit-to-speed-up-functions>`__
-   - SCICO can compute the adjoint of linear operators automatically,
-     which saves time when defining new operators.
-   - SCICO's operator calculus makes code for optimization routines
-     look like the pseudocode in scientific papers.
-   - SCICO provides state-of-the-art optimization routines,
-     including projected gradients
-     and the alternating direction method of multipliers (ADMM)
-     with the flexibility of plug-and-play priors
-     including BM3D :cite:`dabov-2008-image` and DnCNN :cite:`zhang-2017-dncnn` denoisers.
+   - Operators in SCICO are defined using a simple, NumPy-like syntax, yet automatically benefit from GPU/TPU acceleration and `just-in-time compilation <https://jax.readthedocs.io/en/latest/notebooks/quickstart.html#using-jit-to-speed-up-functions>`__.
+   - SCICO can compute the adjoint of linear operators automatically, which saves time when defining new operators.
+   - SCICO's operator calculus makes code for optimization routines look like the pseudocode in scientific papers.
+   - SCICO provides state-of-the-art optimization routines, including projected gradients and the alternating direction method of multipliers (ADMM),
+     with the flexibility of plug-and-play priors :cite:`venkatakrishnan-2013-plugandplay2` with BM3D :cite:`dabov-2008-image` and DnCNN :cite:`zhang-2017-dncnn` denoisers.
 
 
-If you use this library for published work,
-please cite :cite:`scico-2022`
-(see bibtex entry ``scico-2022`` in
-`docs/source/references.bib
-<https://github.com/lanl/scico/blob/main/docs/source/references.bib>`_
-in the source distribution).
+If you use this library for published work, please cite :cite:`scico-2022` (see bibtex entry ``scico-2022`` in `docs/source/references.bib <https://github.com/lanl/scico/blob/main/docs/source/references.bib>`_ in the source distribution).
+
+
+Computational Imaging
+---------------------
+
[…]
