+++ docs/source/overview.rst
@@ -25,128 +25,21 @@ Overview
 
 
 
-`SCICO <https://github.com/lanl/scico>`__ is a Python package for solving the inverse problems that arise in scientific imaging applications.  Its primary focus is providing methods for solving ill-posed inverse problems by using an appropriate prior model of the reconstruction space.  SCICO includes a growing suite of operators, cost functionals, regularizers, and optimization routines that may be combined to solve a wide range of problems and is designed so that it is easy to add new building blocks.
+`Scientific Computational Imaging Code (SCICO) <https://github.com/lanl/scico>`__ is a Python package for solving the inverse problems that arise in scientific imaging applications. Its primary focus is providing methods for solving ill-posed inverse problems by using an appropriate prior model of the reconstruction space. SCICO includes a growing suite of operators, cost functionals, regularizers, and optimization algorithms that may be combined to solve a wide range of problems, and is designed so that it is easy to add new building blocks. When solving a problem, these components are combined in a way that makes code for optimization routines look like the pseudocode in scientific papers. SCICO is built on top of `JAX <https://jax.readthedocs.io/en/latest/>`__ rather than `NumPy <https://numpy.org/>`__, enabling GPU/TPU acceleration, just-in-time compilation, and automatic gradient functionality, which is used to automatically compute the adjoints of linear operators. An example of how to solve a multi-channel tomography problem with SCICO is shown in the figure below.
 
-SCICO's main advantages are:
 
-   - Operators in SCICO are defined using a simple, NumPy-like syntax, yet automatically benefit from GPU/TPU acceleration and `just-in-time compilation <https://jax.readthedocs.io/en/latest/notebooks/quickstart.html#using-jit-to-speed-up-functions>`__.
-   - SCICO can compute the adjoint of linear operators automatically, which saves time when defining new operators.
-   - SCICO's operator calculus makes code for optimization routines look like the pseudocode in scientific papers.
-   - SCICO provides state-of-the-art optimization routines, including projected gradients and the alternating direction method of multipliers (ADMM),
-     with the flexibility of plug-and-play priors :cite:`venkatakrishnan-2013-plugandplay2` with BM3D :cite:`dabov-2008-image` and DnCNN :cite:`zhang-2017-dncnn` denoisers.
+.. image:: /figures/scico-tomo-overview.png
+     :align: center
+     :width: 95%
+     :alt: Solving a multi-channel tomography problem with SCICO.
+
+|
+
+The SCICO source code is available from `GitHub <https://github.com/lanl/scico>`__, and pre-built packages are available from `PyPI <https://github.com/lanl/scico>`__. (Detailed insructions for installing SCICO are available in :ref:`installing`.) It has extensive `online documentation <https://scico.rtfd.io/>`__, including API documentation and usage examples, which can be run online at `Google Colab <https://colab.research.google.com/github/lanl/scico-data/blob/colab/notebooks/index.ipynb>`__ and `binder <https://mybinder.org/v2/gh/lanl/scico-data/binder?labpath=notebooks%2Findex.ipynb>`__.
 
 
 If you use this library for published work, please cite :cite:`scico-2022` (see bibtex entry ``scico-2022`` in `docs/source/references.bib <https://github.com/lanl/scico/blob/main/docs/source/references.bib>`_ in the source distribution).
 
-
-Computational Imaging
----------------------
-
[…]
