+++ docs/source/advantages.rst
@@ -1,0 +1,89 @@
+Why Use SCICO?
+==============
+
+
+.. raw:: html
+
+    <style type='text/css'>
+    div.document ul blockquote {
+       margin-bottom: 8px !important;
+    }
+    div.document li > p {
+       margin-bottom: 4px !important;
+    }
+    div.document ul > li {
+      list-style: square outside !important;
+      margin-left: 1em !important;
+    }
+    section {
+      padding-bottom: 1em;
+    }
+    ul {
+      margin-bottom: 1em;
+    }
+    </style>
+
+
+
+Advantages of JAX-based Design
+------------------------------
+
+The vast majority of scientific computing packages in Python are based on
+[NumPy](https://numpy.org/) and [SciPy](https://scipy.org/). SCICO, in contrast, is based on [JAX](https://jax.readthedocs.io/en/latest/), which provides most of the same features, but with the addition of automatic differentiation, GPU support, and just-in-time (JIT) compilation.
+
+While recent advances in automatic differentiation have primarily been driven by its important role in deep learning, it is also invaluable in a functional minimization framework such as SCICO. The most obvious advantage is allowing the use of gradient-based minimization methods without the need for tedious mathematical derivation of an expression for the gradient. Equally valuable, though, is the ability to automatically compute the adjoint operator of a linear operator, the manual derivation of which is often time-consuming.
+
+GPU support and JIT compilation both offer the potential for significant code acceleration, with the speed gains that can be obtained depending on the algorithm/function to be executed. In many cases, a speed improvement by an order of magnitude or more can be obtained by running the same code on a GPU rather than a CPU, and similar speed gains can sometimes also be obtained via JIT compilation.
+
+The figure below shows timing results obtained on a compute server with an Intel Xeon Gold 6230 CPU and NVIDIA GeForce RTX 2080 Ti GPU.  It is interesting to note that for `FiniteDifference` the GPU provides no acceleration, while JIT provides more than an order of magnitude of speed improvement on both CPU and GPU. For `DFT` and `Convolve`, significant JIT acceleration is limited to the GPU, which also provides significant acceleration over the CPU.
+
+
+.. image:: /figures/jax-timing.png
+     :align: center
+     :width: 95%
+     :alt: Timing results for SCICO operators on CPU and GPU with and without JIT
+
+
+
+Related Packages
+----------------
+
+Many elements of SCICO are partially available in other packages.
+We briefly review them here, highlighting some of the main differences with SCICO.
+
+[GlobalBioIm](https://biomedical-imaging-group.github.io/GlobalBioIm/)
+is similar in structure to SCICO (and a major inspiration for SCICO),
+providing linear operators and solvers for inverse problems in imaging.
+However, it is written in MATLAB and is thus not usable in a completely free environment. It also lacks the automatic adjoint calculation and simple GPU support offered by SCICO.
+
+[PyLops](https://pylops.readthedocs.io) provides a linear operator
+class and many built-in linear operators.
+These operators are compatible with many [SciPy](https://scipy.org/) solvers.
+GPU support is provided via [CuPy](https://cupy.dev),
+which has the disadvantage that switching for a CPU to GPU requires code changes,
+unlike SCICO and [JAX](https://jax.readthedocs.io/en/latest/).
+SCICO is more focused on computational imaging that PyLops
+and has several specialized operators that PyLops does not.
+
[…]
