### 2022-09-05 f19f011
Fix residual md formatting and add missing references
