+++ docs/source/inverse.rst
@@ -24,7 +24,6 @@ Inverse Problems
     </style>
 
 
-
 In traditional imaging, the burden of image formation is placed on physical components, such as a lens, with the resulting image being taken from the sensor with minimal processing. In computational imaging, in contrast, the burden of image formation is shared with or shifted to computation, with the resulting image typically being very different from the measured data. Common examples of computational imaging include demosaicing in consumer cameras, computed tomography and magnetic resonance imaging in medicine, and synthetic aperture radar in remote sensing. This is an active and growing area of research, and many of these problems have common properties that could be supported by shared implementations of solution components.
 
 The goal of SCICO is to provide a general research tool for computational imaging, with a particular focus on scientific imaging applications, which are particularly underrepresented in the existing range of open-source packages in this area. While a number of other packages overlap somewhat in functionality with SCICO, only a few support execution of the same code on both CPU and GPU devices, and we are not aware of any that support just-in-time compilation and automatic gradient computation, which is invaluable in computational imaging. SCICO provides all three of these valuable features by being built on top of `JAX <https://jax.readthedocs.io/en/latest/>`__ rather than `NumPy <https://numpy.org/>`__.
@@ -33,202 +32,334 @@ The goal of SCICO is to provide a general research tool for computational imagin
 The remainder of this section outlines the steps involved in solving an inverse problem, and shows how each concept maps to a component of SCICO. More detail on the main classes involved in setting up and solving an inverse problem can be found in :ref:`classes`.
 
 
-
 Forward Modeling
 ----------------
 
-In order to solve a computational imaging problem we need to know how the image we wish to reconstruct, $\mb{x}$, is related to the data that we can measure, $\mb{y}$.
-This is represented via a model of the measurement process,
+In order to solve a computational imaging problem we need to know how
+the image we wish to reconstruct, :math:`\mathbf{x}`, is related to the
+data that we can measure, :math:`\mathbf{y}`. This is represented via a
+model of the measurement process,
 
-$$\mb{y} = A(\mb{x}) \,.$$
+.. math:: \mathbf{y} = A(\mathbf{x}) \,.
 
-SCICO provides the `Operator` and `LinearOperator` classes, which may be subclassed by users,
-in order to implement the forward operator, $A$.
-It also has several built-in operators, most of which are linear, e.g.,
-finite convolutions, discrete Fourier transforms, optical propagators (see \autoref{fig:optprop}), Abel transforms, and Radon transforms.
-For example,
+SCICO provides the ``Operator`` and ``LinearOperator`` classes, which
+may be subclassed by users, in order to implement the forward
+operator, :math:`A`. It also has several built-in operators, most of
+which are linear, e.g., finite convolutions, discrete Fourier
+transforms, optical propagators, Abel transforms, and Radon
+transforms. For example,
 
-```python
-	input_shape = (512, 512)
-	angles = np.linspace(0, 2 * np.pi, 180, endpoint=False)
-	channels = 512
-	A = scico.linop.radon_svmbir.ParallelBeamProjector(
-		input_shape, angles, channels
-	)
-```
+.. code:: python
+
+       input_shape = (512, 512)
+       angles = np.linspace(0, 2 * np.pi, 180, endpoint=False)
+       channels = 512
+       A = scico.linop.radon_svmbir.ParallelBeamProjector(
+           input_shape, angles, channels)
 
 defines a tomographic projection operator.
 
-
[…]
