+++ docs/source/advantages.rst
@@ -28,14 +28,36 @@ Why Use SCICO?
 Advantages of JAX-based Design
 ------------------------------
 
-The vast majority of scientific computing packages in Python are based on
-[NumPy](https://numpy.org/) and [SciPy](https://scipy.org/). SCICO, in contrast, is based on [JAX](https://jax.readthedocs.io/en/latest/), which provides most of the same features, but with the addition of automatic differentiation, GPU support, and just-in-time (JIT) compilation.
+The vast majority of scientific computing packages in Python are based
+on `NumPy <https://numpy.org/>`__ and `SciPy <https://scipy.org/>`__.
+SCICO, in contrast, is based on
+`JAX <https://jax.readthedocs.io/en/latest/>`__, which provides most of
+the same features, but with the addition of automatic differentiation,
+GPU support, and just-in-time (JIT) compilation.
 
-While recent advances in automatic differentiation have primarily been driven by its important role in deep learning, it is also invaluable in a functional minimization framework such as SCICO. The most obvious advantage is allowing the use of gradient-based minimization methods without the need for tedious mathematical derivation of an expression for the gradient. Equally valuable, though, is the ability to automatically compute the adjoint operator of a linear operator, the manual derivation of which is often time-consuming.
+While recent advances in automatic differentiation have primarily been
+driven by its important role in deep learning, it is also invaluable in
+a functional minimization framework such as SCICO. The most obvious
+advantage is allowing the use of gradient-based minimization methods
+without the need for tedious mathematical derivation of an expression
+for the gradient. Equally valuable, though, is the ability to
+automatically compute the adjoint operator of a linear operator, the
+manual derivation of which is often time-consuming.
 
-GPU support and JIT compilation both offer the potential for significant code acceleration, with the speed gains that can be obtained depending on the algorithm/function to be executed. In many cases, a speed improvement by an order of magnitude or more can be obtained by running the same code on a GPU rather than a CPU, and similar speed gains can sometimes also be obtained via JIT compilation.
+GPU support and JIT compilation both offer the potential for significant
+code acceleration, with the speed gains that can be obtained depending
+on the algorithm/function to be executed. In many cases, a speed
+improvement by an order of magnitude or more can be obtained by running
+the same code on a GPU rather than a CPU, and similar speed gains can
+sometimes also be obtained via JIT compilation.
 
-The figure below shows timing results obtained on a compute server with an Intel Xeon Gold 6230 CPU and NVIDIA GeForce RTX 2080 Ti GPU.  It is interesting to note that for `FiniteDifference` the GPU provides no acceleration, while JIT provides more than an order of magnitude of speed improvement on both CPU and GPU. For `DFT` and `Convolve`, significant JIT acceleration is limited to the GPU, which also provides significant acceleration over the CPU.
+The figure below shows timing results obtained on a compute server with
+an Intel Xeon Gold 6230 CPU and NVIDIA GeForce RTX 2080 Ti GPU. It is
+interesting to note that for ``FiniteDifference`` the GPU provides no
+acceleration, while JIT provides more than an order of magnitude of
+speed improvement on both CPU and GPU. For ``DFT`` and ``Convolve``,
+significant JIT acceleration is limited to the GPU, which also provides
+significant acceleration over the CPU.
 
 
 .. image:: /figures/jax-timing.png
[…]

[…]
