### 2022-09-05 27382df
Work in progress on NonLinearPADMM
