+++ scico/optimize/_nlpadmm.py
@@ -1,0 +1,340 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2022 by SCICO Developers
+# All rights reserved. BSD 3-clause License.
+# This file is part of the SCICO package. Details of the copyright and
+# user license can be found in the 'LICENSE' file distributed with the
+# package.
+
+"""Non-linear Proximal ADMM solver."""
+
+# Needed to annotate a class method that returns the encapsulating class;
+# see https://www.python.org/dev/peps/pep-0563/
+from __future__ import annotations
+
+from typing import Callable, List, Optional, Tuple, Union
+
+import scico.numpy as snp
+from scico.diagnostics import IterationStats
+from scico.functional import Functional
+from scico.numpy import BlockArray
+from scico.numpy.linalg import norm
+from scico.numpy.util import ensure_on_device
+from scico.typing import JaxArray
+from scico.util import Timer
+
+
+class NonLinearPADMM:
+    r"""Non-linear proximal alternating direction method of multipliers.
+
+    |
+
+    Solve an optimization problem of the form
+
+    .. math::
+        \argmin_{\mb{x}} \; f(\mb{x}) + g(\mb{z}) \;
+        \text{such that}\; H(\mb{x}, \mb{z}) = 0 \;,
+
+    where :math:`f` and :math:`g` are instances of :class:`.Functional`,
+    (in most cases :math:`f` will, more specifically be an an instance
+    of :class:`.Loss`), and :math:`H` is a function.
+
+    The optimization problem is solved via a variant of the proximal ADMM
+    algorithm for problems with a non-linear operator constraint
+    :cite:`benning-2016-preconditioned`, consisting of the
+    iterations (see :meth:`step`)
+
+    .. math::
+       \begin{aligned}
+       A^{(k)} &= J_{\mb{x}} H(\mb{x}^{(k)}, \mb{z}^{(k)}) \\
+       \mb{x}^{(k+1)} &= \mathrm{prox}_{\rho^{-1} \mu^{-1} f} \left(
+         \mb{x}^{(k)} - \mu^{-1} (A^{(k)})^T \left(2 \mb{u}^{(k)} -
+         \mb{u}^{(k-1)} \right) \right) \\
+       B^{(k)} &= J_{\mb{z}} H(\mb{x}^{(k+1)}, \mb{z}^{(k)}) \\
+       \mb{z}^{(k+1)} &= \mathrm{prox}_{\rho^{-1} \nu^{-1} g} \left(
+         \mb{z}^{(k)} - \nu^{-1} (B^{(k)})^T \left(
+         H(\mb{x}^{(k+1)}, \mb{z}^{(k)}) + \mb{u}^{(k)} \right) \right) \\
+       \mb{u}^{(k+1)} &=  \mb{u}^{(k)} + H(\mb{x}^{(k+1)},
+         \mb{z}^{(k+1)})  \;.
+       \end{aligned}
+
+    Parameters :math:`\mu` and :math:`\nu` are required to satisfy
+
+    .. math::
+       \mu > \norm{ A^{(k)} }_2^2 \quad \text{and} \quad \nu > \norm{ B^{(k)} }_2^2
+
+    for all :math:`A^{(k)}` and :math:`B^{(k)}`.
+
+
+    Attributes:
+        f (:class:`.Functional`): Functional :math:`f` (usually a
+           :class:`.Loss`).
+        g (:class:`.Functional`): Functional :math:`g`.
+        H (Callable): :math:`H` function.
+        itnum (int): Iteration counter.
+        maxiter (int): Number of linearized ADMM outer-loop iterations.
+        timer (:class:`.Timer`): Iteration timer.
+        rho (scalar): Penalty parameter.
+        mu (scalar): First algorithm parameter.
[…]
