### 2022-09-05 c995d78
Fixed hcb call with device. Solver tests are fixed and should test device addition.
