+++ docs/source/inverse.rst
@@ -42,10 +42,10 @@ model of the measurement process,
 
 .. math:: \mathbf{y} = A(\mathbf{x}) \,.
 
-SCICO provides the ``Operator`` and ``LinearOperator`` classes, which
-may be subclassed by users, in order to implement the forward
-operator, :math:`A`. It also has several built-in operators, most of
-which are linear, e.g., finite convolutions, discrete Fourier
+SCICO provides the :class:`.Operator` and :class:`.LinearOperator`
+classes, which may be subclassed by users, in order to implement the
+forward operator, :math:`A`. It also has several built-in operators,
+most of which are linear, e.g., finite convolutions, discrete Fourier
 transforms, optical propagators, Abel transforms, and Radon
 transforms. For example,
 
@@ -59,43 +59,47 @@ transforms. For example,
 
 defines a tomographic projection operator.
 
-A significant advantage of SCICO being built on top of
-`JAX <https://jax.readthedocs.io/en/latest/>`__ is that the adjoints of
+A significant advantage of SCICO being built on top of `JAX
+<https://jax.readthedocs.io/en/latest/>`__ is that the adjoints of
 linear operators, which can be quite time consuming to implement even
-when the operator itself is straightforward, are computed automatically
-by exploiting the automatic differentation features of
-`JAX <https://jax.readthedocs.io/en/latest/>`__. If ``A`` is a
-``LinearOperator``, then its adjoint is simply ``A.T`` for real
-transforms and ``A.H`` for complex transforms. Likewise, Jacobian-vector
-products can be automatically computed for non-linear operators,
-allowing for simple linearization and gradient calculations.
+when the operator itself is straightforward, are computed
+automatically by exploiting the automatic differentation features of
+`JAX <https://jax.readthedocs.io/en/latest/>`__. If :code:`A` is a
+:class:`.LinearOperator`, then its adjoint is simply :code:`A.T` for
+real transforms and :code:`A.H` for complex transforms. Likewise,
+Jacobian-vector products can be automatically computed for non-linear
+operators, allowing for simple linearization and gradient
+calculations.
 
 SCICO operators can be composed to construct new operators. (If both
 operands are linear, then the result is also linear.) For example, if
-``A`` and ``B`` have been defined as distinct linear operators, then
+:code:`A` and :code:`B` have been defined as distinct linear
+operators, then
 
 .. code:: python
 
        C = B @ A
 
-defines a new linear operator ``C`` that first applies operator ``A``
-and then applies operator ``B`` to the result (i.e. :math:`C = B A` in
-math notation). This operator algebra can be used to build complicated
-forward operators from simpler building blocks.
+defines a new linear operator :code:`C` that first applies operator
+:code:`A` and then applies operator :code:`B` to the result
+(i.e. :math:`C = B A` in math notation). This operator algebra can be
+used to build complicated forward operators from simpler building
+blocks.
 
-SCICO also handles cases where either the image we want to reconstruct,
-:math:`\mathbf{x}`, or its measurements, :math:`\mathbf{y}`, do not fit
-neatly into a multi-dimensional array. This is achieved via
-``BlockArray`` objects, which consist of a ``list`` of multi-dimensional
-array *blocks*. A ``BlockArray`` differs from a ``list`` in that,
[…]
