+++ docs/source/references.bib
@@ -569,7 +569,6 @@
   doi =		 {10.1364/ao.48.006132},
 }
 
-
 @Book {voelz-2011-computational,
   author =	 {Voelz, David},
   title =	 {Computational {F}ourier optics : a {MATLAB}
@@ -579,7 +578,6 @@
   address =	 {Bellingham, Wash},
   isbn =	 9780819482044,
 }
-
 
 @Article {yang-2012-linearized,
   author =	 {Junfeng Yang and Xiaoming Yuan},
@@ -594,7 +592,6 @@
   pages =	 {301--329}
 }
 
-
 @Article {zhang-2017-dncnn,
   author =	 {Kai Zhang and Wangmeng Zuo and Yunjin Chen and Deyu
                   Meng and Lei Zhang},
@@ -608,7 +605,6 @@
   doi =		 {10.1109/TIP.2017.2662206},
   pages =	 {3142--3155}
 }
-
 
 @Article {zhou-2006-adaptive,
   author =	 {Bin Zhou and Li Gao and Yu-Hong Dai},
