+++ docs/source/install.rst
@@ -3,8 +3,15 @@
 Installing SCICO
 ================
 
-SCICO requires Python version 3.8 or later. It has been tested on Linux and macOS, but is not currently supported on Windows due to the limited support for ``jaxlib`` on Windows. However, Windows users can use SCICO via the `Windows Subsystem for Linux <https://docs.microsoft.com/en-us/windows/wsl/about>`_. Guides exist for using WSL with `CPU only <https://docs.microsoft.com/en-us/windows/wsl/install-win10>`_ and with
-`GPU support <https://docs.microsoft.com/en-us/windows/win32/direct3d12/gpu-cuda-in-wsl>`_.
+SCICO requires Python version 3.8 or later. It has been tested on
+Linux and macOS, but is not currently supported on Windows due to the
+limited support for ``jaxlib`` on Windows. However, Windows users can
+use SCICO via the `Windows Subsystem for Linux
+<https://docs.microsoft.com/en-us/windows/wsl/about>`_ (WSL). Guides
+exist for using WSL with `CPU only
+<https://docs.microsoft.com/en-us/windows/wsl/install-win10>`_ and
+with `GPU support
+<https://docs.microsoft.com/en-us/windows/win32/direct3d12/gpu-cuda-in-wsl>`_.
 
 
 From PyPI
