+++ requirements.txt
@@ -2,7 +2,7 @@ numpy>=1.12
 scipy>=0.19.1
 tifffile
 imageio>=2.17
-matplotlib<=3.1.3
+matplotlib
 jaxlib>=0.3.0,<=0.3.15
 jax>=0.3.0,<=0.3.15
 flax>=0.4.0,<=0.4.1
