### 2022-09-07 0ff8adc
More minor docs improvements
