### 2022-09-07 e3ebdc4
Corrected syntax for fun. Added device logic for call_with_device. Adding test cases.
