+++ scico/test/test_solver.py
@@ -179,10 +179,14 @@ def test_minimize(dtype, method):
     assert out.x.shape == x.shape
     np.testing.assert_allclose(out.x.ravel(), expected, rtol=5e-4)
 
-    out = solver.minimize(f, x0=jax.device_put(snp.zeros_like(x)), method=method)
+    devices = jax.devices()
 
-    assert out.x.shape == x.shape
-    np.testing.assert_allclose(out.x.ravel(), expected, rtol=5e-4)
+    print(devices)
+    if len(devices) > 1:
+        out = solver.minimize(f, x0=jax.device_put(snp.zeros_like(x), devices[1]), method=method)
+        # Need to add device test here to compare old device and new device
+        assert out.x.shape == x.shape
+        np.testing.assert_allclose(out.x.ravel(), expected, rtol=5e-4)
 
 
 def test_split_join_array():
@@ -211,7 +215,7 @@ def test_split_join_blockarray():
 
 
 def test_bisect():
-    f = lambda x: x**3
+    f = lambda x: x ** 3
     x, info = solver.bisect(f, -snp.ones((5, 1)), snp.ones((5, 1)), full_output=True)
     assert snp.sum(snp.abs(x)) == 0.0
     assert info["iter"] == 0
@@ -219,14 +223,14 @@ def test_bisect():
     assert snp.max(snp.abs(x)) <= 1e-5
     assert snp.max(snp.abs(f(x))) <= 1e-5
     c, key = random.randn((5, 1), dtype=np.float32)
-    f = lambda x, c: x**3 - c**3
+    f = lambda x, c: x ** 3 - c ** 3
     x = solver.bisect(f, -snp.abs(c) - 1, snp.abs(c) + 1, args=(c,), xtol=1e-5, ftol=1e-5)
     assert snp.max(snp.abs(x - c)) <= 1e-5
     assert snp.max(snp.abs(f(x, c))) <= 1e-5
 
 
 def test_golden():
-    f = lambda x: x**2
+    f = lambda x: x ** 2
     x, info = solver.golden(f, -snp.ones((5, 1)), snp.ones((5, 1)), full_output=True)
     assert snp.max(snp.abs(x)) <= 1e-7
     x = solver.golden(f, -2.0 * snp.ones((5, 3)), snp.ones((5, 3)), xtol=1e-5)
