### 2022-09-07 98d3df7
Address #335 (#336)
* Resolve #335

* Clean up after black

* Improve docs

* Minor docs fix

* More minor docs improvements
