### 2022-09-07 eb62125
Rephrase some text and address #334
