### 2022-09-08 1c19857
Add note on limitations of non-JAX packages
