### 2022-09-08 feb22f2
Bibtex updates
