+++ docs/source/references.bib
@@ -326,15 +326,15 @@
   pages =	 {1872--1876}
 }
 
-@Misc {kamilov-2022-plug,
+@Article {kamilov-2022-plug,
   author =	 {Ulugbek S. Kamilov and Charles A. Bouman and Gregery
                   T. Buzzard and Brendt Wohlberg},
   title =	 {Plug-and-Play Methods for Integrating Physical and
                   Learned Models in Computational Imaging},
-  howpublished = {arXiv},
+  journal =      {IEEE Signal Processing Magazine},
   year =	 2022,
   eprint =	 {arXiv:2203.17061},
-  url =		 {https://arxiv.org/abs/2203.17061}
+  note =         {To appear.}
 }
 
 @Article {liu-2018-first,
@@ -360,7 +360,8 @@
   volume =	 22,
   number =	 1,
   pages =	 {119--133},
-  year =	 2012
+  year =	 2012,
+  doi =          {10.1109/TIP.2012.2210725}
 }
 
 @InProceedings {makinen-2019-exact,
