### 2022-09-09 a5cc09c
Move external package notes to Notes section and restructure that section
