### 2022-09-09 187eb00
Fixed device issues. Added test cases for devices and cleaned up code.
