### 2022-09-10 3633f66
Add flexible partial function creation function
