### 2022-09-11 098769d
Add Function class
