### 2022-09-12 c429a14
Fix docs math for isotropic TV
