+++ examples/scripts/ct_astra_tv_admm.py
@@ -14,7 +14,7 @@ solve a few-view CT reconstruction problem with anisotropic total
 variation (TV) regularization
 
   $$\mathrm{argmin}_{\mathbf{x}} \; (1/2) \| \mathbf{y} - A \mathbf{x}
-  \|_2^2 + \lambda \| C \mathbf{x} \|_1 \;,$$
+  \|_2^2 + \lambda \| C \mathbf{x} \|_{2,1} \;,$$
 
 where $A$ is the Radon transform, $\mathbf{y}$ is the sinogram, $C$ is
 a 2D finite difference operator, and $\mathbf{x}$ is the desired
