+++ examples/scripts/denoise_l1tv_admm.py
@@ -14,7 +14,7 @@ This example demonstrates impulse noise removal via ℓ1 total variation
 minimizing the functional
 
   $$\mathrm{argmin}_{\mathbf{x}} \;  \| \mathbf{y} - \mathbf{x}
-  \|_1 + \lambda \| C \mathbf{x} \|_1 \;,$$
+  \|_1 + \lambda \| C \mathbf{x} \|_{2,1} \;,$$
 
 where $\mathbf{y}$ is the noisy image, $C$ is a 2D finite difference
 operator, and $\mathbf{x}$ is the denoised image.
