### 2022-09-12 22e822b
Correct docs on isotropic/anisotripic
