+++ examples/scripts/deconv_tv_admm.py
@@ -10,7 +10,7 @@ Image Deconvolution (ADMM w/ Total Variation)
 
 This example demonstrates the use of class
 [admm.ADMM](../_autosummary/scico.optimize.rst#scico.optimize.ADMM) to solve
-an image deconvolution problem with anisotropic total variation (TV)
+an image deconvolution problem with isotropic total variation (TV)
 regularization
 
   $$\mathrm{argmin}_{\mathbf{x}} \; \| \mathbf{y} - A \mathbf{x} \|_2^2
