+++ examples/scripts/ct_astra_weighted_tv_admm.py
@@ -5,23 +5,20 @@
 # with the package.
 
 r"""
-Low-Dose CT (ADMM w/ Total Variation)
-=====================================
+TV Regularized Low-Dose CT Reconstruction
+=========================================
 
-This example demonstrates the use of class
-[admm.ADMM](../_autosummary/scico.optimize.rst#scico.optimize.ADMM) to
-solve a low-dose CT reconstruction problem with isotropic total
-variation (TV) regularization
+This example demonstrates solution of a low-dose CT reconstruction problem
+with isotropic total variation (TV) regularization
 
   $$\mathrm{argmin}_{\mathbf{x}} \; (1/2) \| \mathbf{y} - A \mathbf{x}
   \|_W^2 + \lambda \| C \mathbf{x} \|_{2,1} \;,$$
 
-where $A$ is the Radon transform, $\mathbf{y}$ is the sinogram, $C$ is
+where $A$ is the Radon transform, $\mathbf{y}$ is the sinogram, the norm
+weighting $W$ is chosen so that the weighted norm is an approximation to
+the Poisson negative log likelihood :cite:`sauer-1993-local`, $C$ is
 a 2D finite difference operator, and $\mathbf{x}$ is the desired
 image.
-
-The weighted norm is an approximation to the Poisson negative log
-likelihood :cite:`sauer-1993-local`.
 """
 
 import numpy as np
