+++ examples/scripts/ct_svmbir_ppp_bm3d_admm_cg.py
@@ -5,19 +5,18 @@
 # with the package.
 
 """
-CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+CG)
-================================================================
+PPP (with BM3D) CT Reconstruction (ADMM with CG Subproblem Solver)
+==================================================================
 
-This example demonstrates the use of class
-[admm.ADMM](../_autosummary/scico.optimize.rst#scico.optimize.ADMM) to
-solve a tomographic reconstruction problem using the Plug-and-Play Priors
-framework :cite:`venkatakrishnan-2013-plugandplay2`, using BM3D
+This example demonstrates solution of a tomographic reconstruction problem
+using the Plug-and-Play Priors framework
+:cite:`venkatakrishnan-2013-plugandplay2`, using BM3D
 :cite:`dabov-2008-image` as a denoiser and SVMBIR :cite:`svmbir-2020` for
 tomographic projection.
 
-This version uses the data fidelity term as the ADMM f, and thus the
+This version uses the data fidelity term as the ADMM $f$, and thus the
 optimization with respect to the data fidelity uses CG rather than the
-prox of the SVMBIRSquaredL2Loss functional.
+prox of the `SVMBIRSquaredL2Loss` functional.
 """
 
 import numpy as np
