+++ examples/scripts/README.rst
@@ -10,21 +10,21 @@ Computed Tomography
 ^^^^^^^^^^^^^^^^^^^
 
    `ct_abel_tv_admm.py <ct_abel_tv_admm.py>`_
-      Regularized Abel Inversion
-   `ct_astra_pcg.py <ct_astra_pcg.py>`_
-      CT with Preconditioned Conjugate Gradient
+      TV Regularized Abel Inversion
+   `ct_astra_noreg_pcg.py <ct_astra_noreg_pcg.py>`_
+      CT Reconstruction with CG and PCG
    `ct_astra_tv_admm.py <ct_astra_tv_admm.py>`_
-      Few-View CT (ADMM w/ Total Variation)
+      TV Regularized Few-View CT Reconstruction
    `ct_astra_weighted_tv_admm.py <ct_astra_weighted_tv_admm.py>`_
-      Low-Dose CT (ADMM w/ Total Variation)
+      TV Regularized Low-Dose CT Reconstruction
    `ct_svmbir_ppp_bm3d_admm_cg.py <ct_svmbir_ppp_bm3d_admm_cg.py>`_
-      CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+CG)
+      PPP (with BM3D) CT Reconstruction (ADMM with CG Subproblem Solver)
    `ct_svmbir_ppp_bm3d_admm_prox.py <ct_svmbir_ppp_bm3d_admm_prox.py>`_
-      CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+Prox)
+      PPP (with BM3D) CT Reconstruction (ADMM with SVMBIR Fast Prox)
    `ct_fan_svmbir_ppp_bm3d_admm_prox.py <ct_fan_svmbir_ppp_bm3d_admm_prox.py>`_
-      Fan-beam CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+Prox)
+      PPP (with BM3D) Fan-Beam CT Reconstruction
    `ct_svmbir_tv_multi.py <ct_svmbir_tv_multi.py>`_
-      CT Reconstruction with TV Regularization
+      TV Regularized CT Reconstruction (Multiple Algorithms)
 
 
 Deconvolution
@@ -67,11 +67,11 @@ Miscellaneous
    `demosaic_ppp_bm3d_admm.py <demosaic_ppp_bm3d_admm.py>`_
       Image Demosaicing (ADMM Plug-and-Play Priors w/ BM3D)
    `denoise_cplx_tv_pdhg.py <denoise_cplx_tv_pdhg.py>`_
-      Complex Total Variation Denoising (ADMM)
+      Complex Total Variation Denoising (PDHG)
    `denoise_l1tv_admm.py <denoise_l1tv_admm.py>`_
       ℓ1 Total Variation (ADMM)
    `denoise_tv_admm.py <denoise_tv_admm.py>`_
-      Isotropic Total Variation (ADMM)
+      Total Variation Denoising (ADMM)
    `denoise_tv_pgm.py <denoise_tv_pgm.py>`_
       Isotropic Total Variation (Accelerated PGM)
    `denoise_tv_multi.py <denoise_tv_multi.py>`_
@@ -90,11 +90,11 @@ Plug and Play Priors
 ^^^^^^^^^^^^^^^^^^^^
 
    `ct_svmbir_ppp_bm3d_admm_cg.py <ct_svmbir_ppp_bm3d_admm_cg.py>`_
-      CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+CG)
+      PPP (with BM3D) CT Reconstruction (ADMM with CG Subproblem Solver)
[…]
