### 2022-09-13 9a9c053
Move significant functions to a separate module
