### 2022-09-13 e529b89
Bump maximum jax version and some other small changes (#338)
* Bump max jax version

* Docstring fix

* Avoid test failure on GPU device
