### 2022-09-13 c2d99d0
Add script to update notebook markdown without re-run
