+++ docs/source/conf.py
@@ -172,7 +172,17 @@ release = version
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
-exclude_patterns = ["tmp", "*.tmp.*", "*.tmp", "index.ipynb", "exampledepend.rst"]
+exclude_patterns = [
+    "tmp",
+    "*.tmp.*",
+    "*.tmp",
+    "index.ipynb",
+    "exampledepend.rst",
+    "blockarray.rst",
+    "operator.rst",
+    "functional.rst",
+    "optimizer.rst",
+]
 
 # If true, '()' will be appended to :func: etc. cross-reference text.
 add_function_parentheses = False
