+++ examples/scripts/deconv_tv_admm.py
@@ -5,20 +5,18 @@
 # with the package.
 
 r"""
-Image Deconvolution (ADMM w/ Total Variation)
-=============================================
+Image Deconvolution with TV Regularization
+==========================================
 
-This example demonstrates the use of class
-[admm.ADMM](../_autosummary/scico.optimize.rst#scico.optimize.ADMM) to solve
-an image deconvolution problem with isotropic total variation (TV)
-regularization
+This example demonstrates the solution of an image deconvolution problem
+with isotropic total variation (TV) regularization
 
-  $$\mathrm{argmin}_{\mathbf{x}} \; \| \mathbf{y} - A \mathbf{x} \|_2^2
-  + \lambda \| C \mathbf{x} \|_{2,1} \;,$$
+  $$\mathrm{argmin}_{\mathbf{x}} \; (1/2) \| \mathbf{y} - A \mathbf{x}
+  \|_2^2 + \lambda \| C \mathbf{x} \|_{2,1} \;,$$
 
-where $A$ is Toeplitz matrix, $\mathbf{y}$ is the blurred image, $C$
-is a 2D Finite Difference operator, and $\mathbf{x}$ is the desired
-image.
+where $A$ is a convolution operator, $\mathbf{y}$ is the blurred image,
+$C$ is a 2D finite fifference operator, and $\mathbf{x}$ is the
+deconvolved image.
 """
 
 import jax
