+++ docs/joss-paper/paper.bib
@@ -146,15 +146,15 @@
   pages =	 {1872--1876}
 }
 
-@Misc {kamilov-2022-plug,
+@Article {kamilov-2022-plug,
   author =	 {Ulugbek S. Kamilov and Charles A. Bouman and Gregery
                   T. Buzzard and Brendt Wohlberg},
   title =	 {Plug-and-Play Methods for Integrating Physical and
                   Learned Models in Computational Imaging},
-  howpublished = {arXiv},
+  journal =      {IEEE Signal Processing Magazine},
   year =	 2022,
   eprint =	 {arXiv:2203.17061},
-  url =		 {https://arxiv.org/abs/2203.17061}
+  note =         {To appear.}
 }
 
 @Software {bradbury-2018-jax,
