### 2022-09-13 d7ca4ba
Fixed hcb call to properly handle device management for minimize. Solved mypy warnings.
