+++ scico/test/test_solver.py
@@ -185,15 +185,16 @@ def test_minimize(dtype, method):
     # For default device:
     x0 = jax.device_put(snp.zeros_like(x), devices[0])
     out = solver.minimize(f, x0=x0, method=method)
-    assert out.x.device_buffer.device() == devices[0]
+    assert out.x.device() == devices[0]
     assert out.x.shape == x0.shape
     np.testing.assert_allclose(out.x.ravel(), expected, rtol=5e-4)
 
     # If there are more than one device present:
     if len(devices) > 1:
         x0 = jax.device_put(snp.zeros_like(x), devices[1])
+        assert x0.device() == devices[1]
         out = solver.minimize(f, x0=x0, method=method)
-        assert out.x.device_buffer.device() == devices[1]
+        assert out.x.device() == devices[1]
         assert out.x.shape == x0.shape
         np.testing.assert_allclose(out.x.ravel(), expected, rtol=5e-4)
 
