### 2022-09-13 de8b486
Modify index and simplify some titles
