+++ examples/scripts/index.rst
@@ -29,8 +29,8 @@ Deconvolution
    - deconv_microscopy_allchn_tv_admm.py
    - deconv_ppp_bm3d_admm.py
    - deconv_ppp_bm3d_pgm.py
-   - deconv_ppp_bm4d_admm.py
    - deconv_ppp_dncnn_admm.py
+   - deconv_ppp_bm4d_admm.py
 
 
 Sparse Coding
