### 2022-09-13 ad1508e
Fixing dtype of result from hcb float32 vs float64
