### 2022-09-13 8165b56
Fix bug in notebook build functions
