### 2022-09-13 ef3919d
Returned proper reshaped x0 object in hcb
