+++ examples/scripts/README.rst
@@ -17,14 +17,14 @@ Computed Tomography
       TV Regularized Few-View CT Reconstruction
    `ct_astra_weighted_tv_admm.py <ct_astra_weighted_tv_admm.py>`_
       TV Regularized Low-Dose CT Reconstruction
+   `ct_svmbir_tv_multi.py <ct_svmbir_tv_multi.py>`_
+      TV Regularized CT Reconstruction (Multiple Algorithms)
    `ct_svmbir_ppp_bm3d_admm_cg.py <ct_svmbir_ppp_bm3d_admm_cg.py>`_
       PPP (with BM3D) CT Reconstruction (ADMM with CG Subproblem Solver)
    `ct_svmbir_ppp_bm3d_admm_prox.py <ct_svmbir_ppp_bm3d_admm_prox.py>`_
       PPP (with BM3D) CT Reconstruction (ADMM with SVMBIR Fast Prox)
    `ct_fan_svmbir_ppp_bm3d_admm_prox.py <ct_fan_svmbir_ppp_bm3d_admm_prox.py>`_
       PPP (with BM3D) Fan-Beam CT Reconstruction
-   `ct_svmbir_tv_multi.py <ct_svmbir_tv_multi.py>`_
-      TV Regularized CT Reconstruction (Multiple Algorithms)
 
 
 Deconvolution
@@ -68,8 +68,6 @@ Miscellaneous
       PPP (with BM3D) Image Demosaicing
    `superres_ppp_dncnn_admm.py <superres_ppp_dncnn_admm.py>`_
       PPP (with DnCNN) Image Superresolution
-   `denoise_cplx_tv_pdhg.py <denoise_cplx_tv_pdhg.py>`_
-      Complex Total Variation Denoising
    `denoise_l1tv_admm.py <denoise_l1tv_admm.py>`_
       ℓ1 Total Variation Denoising
    `denoise_tv_admm.py <denoise_tv_admm.py>`_
@@ -78,6 +76,8 @@ Miscellaneous
       Total Variation Denoising with Constraint (APGM)
    `denoise_tv_multi.py <denoise_tv_multi.py>`_
       Comparison of Optimization Algorithms for Total Variation Denoising
+   `denoise_cplx_tv_pdhg.py <denoise_cplx_tv_pdhg.py>`_
+      Complex Total Variation Denoising
    `video_rpca_admm.py <video_rpca_admm.py>`_
       Video Decomposition via Robust PCA
 
@@ -99,10 +99,10 @@ Plug and Play Priors
       PPP (with BM3D) Image Deconvolution (ADMM Solver)
    `deconv_ppp_bm3d_pgm.py <deconv_ppp_bm3d_pgm.py>`_
       PPP (with BM3D) Image Deconvolution (APGM Solver)
-   `deconv_ppp_bm4d_admm.py <deconv_ppp_bm4d_admm.py>`_
-      PPP (with BM4D) Volume Deconvolution
    `deconv_ppp_dncnn_admm.py <deconv_ppp_dncnn_admm.py>`_
       PPP (with DnCNN) Image Deconvolution
+   `deconv_ppp_bm4d_admm.py <deconv_ppp_bm4d_admm.py>`_
+      PPP (with BM4D) Volume Deconvolution
    `demosaic_ppp_bm3d_admm.py <demosaic_ppp_bm3d_admm.py>`_
       PPP (with BM3D) Image Demosaicing
    `superres_ppp_dncnn_admm.py <superres_ppp_dncnn_admm.py>`_
@@ -122,16 +122,14 @@ Total Variation
       TV Regularized CT Reconstruction (Multiple Algorithms)
    `deconv_circ_tv_admm.py <deconv_circ_tv_admm.py>`_
[…]
