+++ docs/source/examples.rst
@@ -26,10 +26,10 @@ Computed Tomography
    examples/ct_astra_noreg_pcg
    examples/ct_astra_tv_admm
    examples/ct_astra_weighted_tv_admm
+   examples/ct_svmbir_tv_multi
    examples/ct_svmbir_ppp_bm3d_admm_cg
    examples/ct_svmbir_ppp_bm3d_admm_prox
    examples/ct_fan_svmbir_ppp_bm3d_admm_prox
-   examples/ct_svmbir_tv_multi
 
 
 Deconvolution
@@ -68,11 +68,11 @@ Miscellaneous
 
    examples/demosaic_ppp_bm3d_admm
    examples/superres_ppp_dncnn_admm
-   examples/denoise_cplx_tv_pdhg
    examples/denoise_l1tv_admm
    examples/denoise_tv_admm
    examples/denoise_tv_pgm
    examples/denoise_tv_multi
+   examples/denoise_cplx_tv_pdhg
    examples/video_rpca_admm
 
 
@@ -94,8 +94,8 @@ Plug and Play Priors
    examples/ct_fan_svmbir_ppp_bm3d_admm_prox
    examples/deconv_ppp_bm3d_admm
    examples/deconv_ppp_bm3d_pgm
-   examples/deconv_ppp_bm4d_admm
    examples/deconv_ppp_dncnn_admm
+   examples/deconv_ppp_bm4d_admm
    examples/demosaic_ppp_bm3d_admm
    examples/superres_ppp_dncnn_admm
 
@@ -111,15 +111,15 @@ Total Variation
    examples/ct_astra_weighted_tv_admm
    examples/ct_svmbir_tv_multi
    examples/deconv_circ_tv_admm
-   examples/deconv_microscopy_tv_admm
-   examples/deconv_microscopy_allchn_tv_admm
    examples/deconv_tv_admm
    examples/deconv_tv_admm_tune
-   examples/denoise_cplx_tv_pdhg
+   examples/deconv_microscopy_tv_admm
+   examples/deconv_microscopy_allchn_tv_admm
    examples/denoise_l1tv_admm
    examples/denoise_tv_admm
    examples/denoise_tv_pgm
    examples/denoise_tv_multi
+   examples/denoise_cplx_tv_pdhg
 
 
 Sparsity
@@ -150,24 +150,24 @@ ADMM
    examples/ct_abel_tv_admm
    examples/ct_astra_tv_admm
    examples/ct_astra_weighted_tv_admm
+   examples/ct_svmbir_tv_multi
    examples/ct_svmbir_ppp_bm3d_admm_cg
    examples/ct_svmbir_ppp_bm3d_admm_prox
    examples/ct_fan_svmbir_ppp_bm3d_admm_prox
-   examples/ct_svmbir_tv_multi
    examples/deconv_circ_tv_admm
+   examples/deconv_tv_admm
+   examples/deconv_tv_admm_tune
    examples/deconv_microscopy_tv_admm
    examples/deconv_microscopy_allchn_tv_admm
    examples/deconv_ppp_bm3d_admm
-   examples/deconv_ppp_bm4d_admm
    examples/deconv_ppp_dncnn_admm
-   examples/deconv_tv_admm
-   examples/deconv_tv_admm_tune
+   examples/deconv_ppp_bm4d_admm
+   examples/sparsecode_admm
    examples/demosaic_ppp_bm3d_admm
+   examples/superres_ppp_dncnn_admm
    examples/denoise_l1tv_admm
    examples/denoise_tv_admm
    examples/denoise_tv_multi
-   examples/sparsecode_admm
[…]

[…]
