### 2022-09-14 b128967
Changes required by new blockarray implementation
