+++ scico/linop/_grad.py
@@ -17,7 +17,7 @@ from typing import Optional, Tuple, Union
 import numpy as np
 
 import scico.numpy as snp
-from scico.blockarray import BlockArray
+from scico.numpy import BlockArray
 from scico.typing import DType, JaxArray, Shape
 
 from ._linop import LinearOperator
@@ -130,7 +130,7 @@ class ProjectedGradient(LinearOperator):
             if len(self.axes) == 1:
                 return grad
             else:
-                return BlockArray.array(grad)
+                return snp.blockarray(grad)
         else:
             # If coord attribute is not None, return gradients projected onto specified local
             # coordinate systems.
@@ -138,7 +138,7 @@ class ProjectedGradient(LinearOperator):
             if len(self.coord) == 1:
                 return projgrad[0]
             else:
-                return BlockArray.array(projgrad)
+                return snp.blockarray(projgrad)
 
 
 class PolarGradient(ProjectedGradient):
@@ -213,9 +213,9 @@ class PolarGradient(ProjectedGradient):
             theta = snp.expand_dims(theta, single)
         coord = []
         if angular:
-            coord.append(BlockArray.array([-snp.cos(theta), snp.sin(theta)]))
+            coord.append(snp.blockarray([-snp.cos(theta), snp.sin(theta)]))
         if radial:
-            coord.append(BlockArray.array([snp.sin(theta), snp.cos(theta)]))
+            coord.append(snp.blockarray([snp.sin(theta), snp.cos(theta)]))
         super().__init__(
             input_shape=input_shape,
             input_dtype=input_dtype,
@@ -309,11 +309,11 @@ class CylindricalGradient(ProjectedGradient):
             theta = snp.expand_dims(theta, single)
         coord = []
         if angular:
-            coord.append(BlockArray.array([-snp.cos(theta), snp.sin(theta), np.array([0])]))
+            coord.append(snp.blockarray([-snp.cos(theta), snp.sin(theta), np.array([0.0])]))
         if radial:
-            coord.append(BlockArray.array([snp.sin(theta), snp.cos(theta), np.array([0])]))
+            coord.append(snp.blockarray([snp.sin(theta), snp.cos(theta), np.array([0.0])]))
         if axial:
-            coord.append(BlockArray.array([np.array([0]), np.array([0]), np.array([1])]))
+            coord.append(snp.blockarray([np.array([0.0]), np.array([0.0]), np.array([1.0])]))
         super().__init__(
             input_shape=input_shape,
             input_dtype=input_dtype,
@@ -393,7 +393,7 @@ class SphericalGradient(ProjectedGradient):
         end = snp.array(axes_shape) - center
         g0, g1, g2 = np.ogrid[-center[0] : end[0], -center[1] : end[1], -center[2] : end[2]]
         theta = snp.arctan2(g1, g0)
-        phi = snp.arctan2(np.sqrt(g0 ** 2 + g1 ** 2), g2)
+        phi = snp.arctan2(np.sqrt(g0**2 + g1**2), g2)
         # Re-order theta and phi axes in case indices in axes parameter are not in
         # increasing order.
         axis_order = np.argsort(axes)
@@ -407,16 +407,16 @@ class SphericalGradient(ProjectedGradient):
             phi = snp.expand_dims(phi, single)
         coord = []
         if azimuthal:
-            coord.append(BlockArray.array([snp.sin(theta), -snp.cos(theta), np.array([0])]))
+            coord.append(snp.blockarray([snp.sin(theta), -snp.cos(theta), np.array([0.0])]))
         if polar:
             coord.append(
-                BlockArray.array(
+                snp.blockarray(
[…]
