+++ scico/linop/__init__.py
@@ -15,6 +15,12 @@ from ._dft import DFT
 from ._diag import Diagonal, Identity
 from ._diff import FiniteDifference, SingleAxisFiniteDifference
 from ._func import Crop, Pad, Reshape, Slice, Sum, Transpose, linop_from_function
+from ._grad import (
+    CylindricalGradient,
+    PolarGradient,
+    ProjectedGradient,
+    SphericalGradient,
+)
 from ._linop import ComposedLinearOperator, LinearOperator
 from ._matrix import MatrixOperator
 from ._stack import DiagonalStack, VerticalStack
