### 2022-09-14 b09c26b
Resolve some components of #228 (#340)
* Fix docs math for isotropic TV

* Correct docs on isotropic/anisotripic

* Correct example title

* Improve some titles and header descriptions

* Minor edit

* Rename example

* Correct example title

* Clean up titles and header descriptions

* Update example index

* Move significant functions to a separate module

* Fix imports

* Add script to update notebook markdown without re-run

* Update readme

* Avoid duplicate label warnings

* Re-arrange examples in index

* Clean up titles and header descriptions

* Updated example index

* Update submodu
[…]
