+++ examples/scripts/deconv_microscopy_tv_admm.py
@@ -18,7 +18,7 @@ The deconvolution problem is solved using class
 solve an image deconvolution problem with isotropic total variation (TV)
 regularization
 
-  $$\mathrm{argmin}_{\mathbf{x}} \; \| M (\mathbf{y} - A \mathbf{x})
+  $$\mathrm{argmin}_{\mathbf{x}} \; (1/2) \| M (\mathbf{y} - A \mathbf{x})
   \|_2^2 + \lambda \| C \mathbf{x} \|_{2,1} +
   \iota_{\mathrm{NN}}(\mathbf{x}) \;,$$
 
