+++ examples/scripts/README.rst
@@ -10,74 +10,74 @@ Computed Tomography
 ^^^^^^^^^^^^^^^^^^^
 
    `ct_abel_tv_admm.py <ct_abel_tv_admm.py>`_
-      Regularized Abel Inversion
-   `ct_astra_pcg.py <ct_astra_pcg.py>`_
-      CT with Preconditioned Conjugate Gradient
+      TV-Regularized Abel Inversion
+   `ct_astra_noreg_pcg.py <ct_astra_noreg_pcg.py>`_
+      CT Reconstruction with CG and PCG
    `ct_astra_tv_admm.py <ct_astra_tv_admm.py>`_
-      Few-View CT (ADMM w/ Total Variation)
+      TV-Regularized Sparse-View CT Reconstruction
    `ct_astra_weighted_tv_admm.py <ct_astra_weighted_tv_admm.py>`_
-      Low-Dose CT (ADMM w/ Total Variation)
-   `ct_svmbir_ppp_bm3d_admm_cg.py <ct_svmbir_ppp_bm3d_admm_cg.py>`_
-      CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+CG)
-   `ct_svmbir_ppp_bm3d_admm_prox.py <ct_svmbir_ppp_bm3d_admm_prox.py>`_
-      CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+Prox)
-   `ct_fan_svmbir_ppp_bm3d_admm_prox.py <ct_fan_svmbir_ppp_bm3d_admm_prox.py>`_
-      Fan-beam CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+Prox)
+      TV-Regularized Low-Dose CT Reconstruction
    `ct_svmbir_tv_multi.py <ct_svmbir_tv_multi.py>`_
-      CT Reconstruction with TV Regularization
+      TV-Regularized CT Reconstruction (Multiple Algorithms)
+   `ct_svmbir_ppp_bm3d_admm_cg.py <ct_svmbir_ppp_bm3d_admm_cg.py>`_
+      PPP (with BM3D) CT Reconstruction (ADMM with CG Subproblem Solver)
+   `ct_svmbir_ppp_bm3d_admm_prox.py <ct_svmbir_ppp_bm3d_admm_prox.py>`_
+      PPP (with BM3D) CT Reconstruction (ADMM with Fast SVMBIR Prox)
+   `ct_fan_svmbir_ppp_bm3d_admm_prox.py <ct_fan_svmbir_ppp_bm3d_admm_prox.py>`_
+      PPP (with BM3D) Fan-Beam CT Reconstruction
 
 
 Deconvolution
 ^^^^^^^^^^^^^
 
    `deconv_circ_tv_admm.py <deconv_circ_tv_admm.py>`_
-      Image Deconvolution (ADMM w/ Total Variation and Circulant Blur)
+      Circulant Blur Image Deconvolution with TV Regularization
+   `deconv_tv_admm.py <deconv_tv_admm.py>`_
+      Image Deconvolution with TV Regularization
+   `deconv_tv_admm_tune.py <deconv_tv_admm_tune.py>`_
+      Image Deconvolution Parameter Tuning
    `deconv_microscopy_tv_admm.py <deconv_microscopy_tv_admm.py>`_
       Deconvolution Microscopy (Single Channel)
    `deconv_microscopy_allchn_tv_admm.py <deconv_microscopy_allchn_tv_admm.py>`_
       Deconvolution Microscopy (All Channels)
    `deconv_ppp_bm3d_admm.py <deconv_ppp_bm3d_admm.py>`_
-      Image Deconvolution (ADMM Plug-and-Play Priors w/ BM3D)
[…]
