### 2022-09-15 33e43bc
Improve conda env creation script
