### 2022-09-15 f26107a
Docs edits to indicate support for python 3.7
