### 2022-09-16 3c2cade
Add support for both fan beam geometries
