### 2022-09-16 b848baf
Improve option checking
