### 2022-09-16 584f57d
Use valid geometry in example script
