### 2022-09-16 f1add83
Bump numpy and scipy version requirements
