+++ requirements.txt
@@ -1,5 +1,5 @@
-numpy>=1.12
-scipy>=0.19.1
+numpy>=1.20.0
+scipy>=1.6.0
 tifffile
 imageio>=2.17
 matplotlib
