### 2022-09-19 2bfac94
Fix bug causing failure in wayland session
