### 2022-09-19 9d242c6
Moved bm3d and bm4d to examples requirements
