### 2022-09-19 4895b33
Add warnings on bm3d/bm4d
