+++ docs/source/exampledepend.rst
@@ -13,3 +13,8 @@ which should be installed via conda:
       pip install -r examples/examples_requirements.txt # Installs other example requirements
 
 The dependencies can also be installed individually as required.
+
+Users are advised to read the warning in :mod:`scico.denoiser` before installing `bm3d <https://pypi.org/project/bm3d>`__ or
+`bm4d <https://pypi.org/project/bm4d>`__.
+
+|
