### 2022-09-19 e752d5e
Add installation of bm3d and bm4d, no longer in requirements.txt
