### 2022-09-19 888be06
Correct installation of mamba
