### 2022-09-19 e649b42
Various changes (#341)
* Improve conda env creation script

* Docs edits to indicate support for python 3.7

* Add note on failure under wayland

* Address some deprecation warnings from ray

* Update to match requirements

* Add support for both fan beam geometries

* Improve option checking

* Use valid geometry in example script

* Update submodule

* Add ray version requirement

* Deal with ray deprecation warnings

* Bump numpy and scipy version requirements

* Update change summary

* Fix sphinx build

* Fix bug causing failure in wayland session

* minor style changes

* Correct install
[…]
