### 2022-09-20 53d9d9e
Revert changes for bm3d: culprit appears to be bm4d only
