+++ examples/examples_requirements.txt
@@ -1,5 +1,4 @@
 astra-toolbox
-bm3d>=3.0.9
 bm4d>=4.0.0
 colour_demosaicing
 xdesign>=0.5.5
