+++ requirements.txt
@@ -6,5 +6,6 @@ matplotlib
 jaxlib>=0.3.0,<=0.3.15
 jax>=0.3.0,<=0.3.17
 flax>=0.4.0
+bm3d>=3.0.9
 svmbir>=0.3.0
 pyabel>=0.8.5
