### 2022-09-20 391f0e5
Mitigate impact of #342 (#343)
* Moved bm3d and bm4d to examples requirements

* Add warnings on bm3d/bm4d

* Add installation of bm3d and bm4d, no longer in requirements.txt

* Revert changes for bm3d: culprit appears to be bm4d only

* Add warning if bm4d impact on subnormals detected

* Fix overlength line

* Fix oversight in docs changes
