### 2022-09-22 96564c1
Avoid duplicate label warnings
