### 2022-09-22 0066ffa
Changes for 0.0.3 release (#346)
* Change development status

* Set release date and version number

* Bump date

* Avoid duplicate label warnings
