+++ scico/__init__.py
@@ -8,7 +8,7 @@
 solving the inverse problems that arise in scientific imaging applications.
 """
 
-__version__ = "0.0.3.dev0"
+__version__ = "0.0.3"
 
 import sys
 
