+++ CHANGES.rst
@@ -3,6 +3,13 @@ SCICO Release Notes
 ===================
 
 
+Version 0.0.4   (unreleased)
+----------------------------
+
+• No changes yet.
+
+
+
 Version 0.0.3   (2022-09-21)
 ----------------------------
 
