+++ examples/scripts/index.rst
@@ -10,28 +10,27 @@ Computed Tomography
 ^^^^^^^^^^^^^^^^^^^
 
    - ct_abel_tv_admm.py
-   - ct_astra_pcg.py
+   - ct_astra_noreg_pcg.py
    - ct_astra_tv_admm.py
    - ct_astra_weighted_tv_admm.py
+   - ct_svmbir_tv_multi.py
    - ct_svmbir_ppp_bm3d_admm_cg.py
    - ct_svmbir_ppp_bm3d_admm_prox.py
    - ct_fan_svmbir_ppp_bm3d_admm_prox.py
-   - ct_svmbir_tv_multi.py
-   - ct_astra_unet_train_foam2.py
 
 
 Deconvolution
 ^^^^^^^^^^^^^
 
    - deconv_circ_tv_admm.py
+   - deconv_tv_admm.py
+   - deconv_tv_admm_tune.py
    - deconv_microscopy_tv_admm.py
    - deconv_microscopy_allchn_tv_admm.py
    - deconv_ppp_bm3d_admm.py
-   - deconv_ppp_bm4d_admm.py
    - deconv_ppp_bm3d_pgm.py
    - deconv_ppp_dncnn_admm.py
-   - deconv_tv_admm.py
-   - deconv_tv_admm_tune.py
+   - deconv_ppp_bm4d_admm.py
 
 
 Sparse Coding
@@ -46,16 +45,13 @@ Miscellaneous
 ^^^^^^^^^^^^^
 
    - demosaic_ppp_bm3d_admm.py
-   - denoise_cplx_tv_pdhg.py
+   - superres_ppp_dncnn_admm.py
    - denoise_l1tv_admm.py
    - denoise_tv_admm.py
    - denoise_tv_pgm.py
    - denoise_tv_multi.py
-   - superres_ppp_dncnn_admm.py
+   - denoise_cplx_tv_pdhg.py
    - video_rpca_admm.py
-   - denoise_datagen_bsds.py
-   - deconv_datagen_bsds.py
-   - ct_astra_datagen_foam2.py
 
 
 
@@ -69,9 +65,9 @@ Plug and Play Priors
    - ct_svmbir_ppp_bm3d_admm_prox.py
    - ct_fan_svmbir_ppp_bm3d_admm_prox.py
    - deconv_ppp_bm3d_admm.py
-   - deconv_ppp_bm4d_admm.py
    - deconv_ppp_bm3d_pgm.py
    - deconv_ppp_dncnn_admm.py
+   - deconv_ppp_bm4d_admm.py
    - demosaic_ppp_bm3d_admm.py
    - superres_ppp_dncnn_admm.py
 
@@ -84,15 +80,15 @@ Total Variation
    - ct_astra_weighted_tv_admm.py
    - ct_svmbir_tv_multi.py
    - deconv_circ_tv_admm.py
-   - deconv_microscopy_tv_admm.py
-   - deconv_microscopy_allchn_tv_admm.py
    - deconv_tv_admm.py
    - deconv_tv_admm_tune.py
-   - denoise_cplx_tv_pdhg.py
+   - deconv_microscopy_tv_admm.py
+   - deconv_microscopy_allchn_tv_admm.py
    - denoise_l1tv_admm.py
    - denoise_tv_admm.py
    - denoise_tv_pgm.py
    - denoise_tv_multi.py
+   - denoise_cplx_tv_pdhg.py
 
 
 Sparsity
@@ -102,15 +98,6 @@ Sparsity
    - sparsecode_pgm.py
    - sparsecode_poisson_pgm.py
    - video_rpca_admm.py
-
-
-Neural Networks
-^^^^^^^^^^^^^^^
-
[…]
