+++ examples/scripts/README.rst
@@ -11,28 +11,31 @@ Computed Tomography
 
    `ct_abel_tv_admm.py <ct_abel_tv_admm.py>`_
       Regularized Abel Inversion
-   `ct_astra_pcg.py <ct_astra_pcg.py>`_
-      CT with Preconditioned Conjugate Gradient
+   `ct_astra_noreg_pcg.py <ct_astra_noreg_pcg.py>`_
+      CT Reconstruction with CG and PCG
    `ct_astra_tv_admm.py <ct_astra_tv_admm.py>`_
       Few-View CT (ADMM w/ Total Variation)
    `ct_astra_weighted_tv_admm.py <ct_astra_weighted_tv_admm.py>`_
       Low-Dose CT (ADMM w/ Total Variation)
+   `ct_svmbir_tv_multi.py <ct_svmbir_tv_multi.py>`_
+      CT Reconstruction with TV Regularization
    `ct_svmbir_ppp_bm3d_admm_cg.py <ct_svmbir_ppp_bm3d_admm_cg.py>`_
       CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+CG)
    `ct_svmbir_ppp_bm3d_admm_prox.py <ct_svmbir_ppp_bm3d_admm_prox.py>`_
       CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+Prox)
    `ct_fan_svmbir_ppp_bm3d_admm_prox.py <ct_fan_svmbir_ppp_bm3d_admm_prox.py>`_
       Fan-beam CT Reconstruction (ADMM Plug-and-Play Priors w/ BM3D, SVMBIR+Prox)
-   `ct_svmbir_tv_multi.py <ct_svmbir_tv_multi.py>`_
-      CT Reconstruction with TV Regularization
-   `ct_astra_train_unet.py <ct_astra_train_unet.py>`_
-      Training of UNet for CT Reconstruction via Denoising of FBP
+
 
 Deconvolution
 ^^^^^^^^^^^^^
 
    `deconv_circ_tv_admm.py <deconv_circ_tv_admm.py>`_
       Image Deconvolution (ADMM w/ Total Variation and Circulant Blur)
+   `deconv_tv_admm.py <deconv_tv_admm.py>`_
+      Image Deconvolution (ADMM w/ Total Variation)
+   `deconv_tv_admm_tune.py <deconv_tv_admm_tune.py>`_
+      Image Deconvolution Parameter Tuning
    `deconv_microscopy_tv_admm.py <deconv_microscopy_tv_admm.py>`_
       Deconvolution Microscopy (Single Channel)
    `deconv_microscopy_allchn_tv_admm.py <deconv_microscopy_allchn_tv_admm.py>`_
@@ -43,10 +46,8 @@ Deconvolution
       Image Deconvolution (PGM Plug-and-Play Priors w/ BM3D)
    `deconv_ppp_dncnn_admm.py <deconv_ppp_dncnn_admm.py>`_
       Image Deconvolution (ADMM Plug-and-Play Priors w/ DnCNN)
-   `deconv_tv_admm.py <deconv_tv_admm.py>`_
-      Image Deconvolution (ADMM w/ Total Variation)
-   `deconv_tv_admm_tune.py <deconv_tv_admm_tune.py>`_
-      Image Deconvolution Parameter Tuning
+   `deconv_ppp_bm4d_admm.py <deconv_ppp_bm4d_admm.py>`_
+      Image Deconvolution (ADMM Plug-and-Play Priors w/ BM4D)
 
 
 Sparse Coding
@@ -65,8 +66,8 @@ Miscellaneous
 
    `demosaic_ppp_bm3d_admm.py <demosaic_ppp_bm3d_admm.py>`_
[…]

[…]
