### 2022-09-28 eba33be
Remove deprecated ray initialization option
