### 2022-09-28 4f3c335
Move jax.tree_map calls to jax.tree_util.tree_map
