### 2022-10-03 18e81da
Remove instances in Flax models where both setup and compact modes were used
