### 2022-10-03 cad3d11
Move all declarations to setup method when used in Flax models
