### 2022-10-03 0c9ad60
Copy ModelParamTraversal from Flax (traverse_util) since marked there for deprecation and removal
