### 2022-10-03 5d3f7c0
Use ModelParamTraversal from train.utils
