### 2022-10-03 23158cc
Add adpated ModelParamTraversal tests
