+++ docs/source/references.bib
@@ -617,3 +617,15 @@
   doi =		 {10.1007/s10589-006-6446-0},
   pages =	 {69--86}
 }
+
+@article{zhang-2021-plug,
+  author =  {Zhang, Kai and Li, Yawei and Zuo, Wangmeng and Zhang,
+                Lei and Van Gool, Luc and Timofte, Radu},
+  title = {Plug-and-Play Image Restoration With Deep Denoiser Prior},
+  journal =  {IEEE Transactions on Pattern Analysis and Machine Intelligence},
+  year =  2022,
+  volume =  44,
+  number =  10,
+  doi = {10.1109/TPAMI.2021.3088914},
+  pages = {6360-6376}
+}
