### 2022-10-04 c943c54
Restore test_flax for ray init in flax examples
