### 2022-10-04 c2a5999
Add comment to data generation example
