### 2022-10-05 #4722 @lucaferranti
Apologies for the huge delay, pretty busy period. I have now allocated Saturday 8th morning (European time) to this.

Small side-comment: when I previously reviewed to JOSS, I was put as assignee on the review issue. This was very helpful, because I generally start and end my workday by checking what is assigned to me on github and so I did not forget about it (it also made finding the issue easier). Has something changed? Personally, I think it would be helpful
