+++ docs/source/examples.rst
@@ -49,6 +49,22 @@ Deconvolution
    examples/deconv_ppp_bm4d_admm
 
 
+Machine Learning
+^^^^^^^^^^^^^^^^
+
+.. toctree::
+   :maxdepth: 1
+
+   examples/ct_astra_datagen_foam2
+   examples/ct_astra_unet_train_foam2
+   examples/deconv_datagen_bsds
+   examples/deconv_datagen_foam
+   examples/deconv_modl_train_foam
+   examples/deconv_odp_train_foam
+   examples/denoise_datagen_bsds
+   examples/denoise_dncnn_train_bsds
+
+
 Sparse Coding
 ^^^^^^^^^^^^^
 
