### 2022-10-09 5960624
Add function supporting jvp with Hermitian transpose
