+++ scico/__init__.py
@@ -15,7 +15,7 @@ import sys
 from . import _python37  # python 3.7 compatibility
 
 # isort: off
-from ._autograd import grad, jacrev, linear_adjoint, value_and_grad
+from ._autograd import grad, jacrev, linear_adjoint, value_and_grad, jhvp
 
 import jax, jaxlib
 
@@ -41,6 +41,7 @@ __all__ = [
     "value_and_grad",
     "linear_adjoint",
     "vjp",
+    "jhvp",
     "jvp",
     "jacfwd",
     "jacrev",
