### 2022-10-09 3abb5ab
Improve mathjax handling
