+++ docs/source/contributing.rst
@@ -16,11 +16,17 @@ Contributing
       list-style: square outside !important;
       margin-left: 1em !important;
     }
+    div.highlight-default.notranslate {
+      margin-bottom: 8px !important;
+    }
     section {
       padding-bottom: 1em;
     }
     ul {
       margin-bottom: 1em;
+    }
+    p {
+      margin-bottom: 8px !important;
     }
     </style>
 
@@ -43,13 +49,11 @@ Installing a Development Version
 
       git clone --recurse-submodules git@github.com:<username>/scico.git
 
-
 4. Create a conda environment using Python 3.7 or later, e.g.:
 
    ::
 
       conda create -n scico python=3.9
-
 
 5. Activate the created conda virtual environment:
 
@@ -57,20 +61,17 @@ Installing a Development Version
 
       conda activate scico
 
-
 6. Change directory to the root of the cloned repository:
 
    ::
 
       cd scico
 
-
 7. Add the ``scico`` repo as an upstream remote to sync your changes:
 
    ::
 
       git remote add upstream https://www.github.com/lanl/scico
-
 
 8. After adding the upstream, the recommended way to install SCICO and its dependencies is via pip:
 
@@ -81,10 +82,8 @@ Installing a Development Version
       pip install -r docs/docs_requirements.txt # Installs documentation requirements
       pip install -e .  # Installs SCICO from the current directory in editable mode
 
-
    For installing dependencies related to the examples please see :ref:`example_notebooks`.
    Installing these are neccessary for the successfull running of the tests.
-
 
 9. The SCICO project uses the `black <https://black.readthedocs.io/en/stable/>`_,
    `isort <https://pypi.org/project/isort/>`_ and `pylint <https://pylint.pycqa.org/en/latest/>`_
@@ -99,7 +98,6 @@ Installing a Development Version
    <https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-pkgs.html#preventing-packages-from-updating-pinning>`__
    of `black <https://black.readthedocs.io/en/stable/>`_ to the version
    number specified in ``dev_requirements.txt``.
-
 
 10. For testing see `Tests`_.
 
@@ -142,13 +140,11 @@ A feature development workflow might look like this:
 
 1. Follow the instructions in `Installing a Development Version`_.
 
-
 2. Sync with the upstream repository:
 
    ::
 
       git pull --rebase origin main --recurse-submodules
-
 
 3. Create a branch to develop from:
 
@@ -156,9 +152,7 @@ A feature development workflow might look like this:
 
       git checkout -b <username>/<brief-description>
 
-
 4. Make your desired changes.
-
 
 5. Run the test suite:
 
@@ -171,7 +165,6 @@ A feature development workflow might look like this:
    ::
 
       pytest scico/test/test_blockarray.py
-
 
 6. When you are finished making changes, create a new commit:
 
@@ -192,13 +185,11 @@ A feature development workflow might look like this:
       git fetch upstream
       git rebase upstream/main
 
-
 8. Push your development upstream:
 
    ::
 
       git push --set-upstream origin <username>/<brief-description>
-
 
 9. Create a new pull request to the ``main`` branch; see `the GitHub instructions <https://docs.github.com/en/github/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request>`_.
 
@@ -215,7 +206,6 @@ The following steps show how to add new data, ``new_data.npz``, to the packaged 
 
 
 1. Create new branches in the main ``scico`` repository as well as in the submodule corresponding to the ``scico-data`` repository (which can be achieved by following the usual branch creation procedure after changing the current directory to ``scico/data``).
-
 
[…]
