### 2022-10-10 18b758d
Exception message consistency (#352)
