+++ scico/optimize/_ladmm.py
@@ -193,7 +193,7 @@ class LinearizedADMM:
             scalar: Value of the objective function.
         """
         if (x is None) != (z is None):
-            raise ValueError("Both or neither of x and z must be supplied")
+            raise ValueError("Both or neither of x and z must be supplied.")
         if x is None:
             x = self.x
             z = self.z

[…]
