### 2022-10-17 990ef64
Indicator documentation notation
Make class docstrings agree with prox docstrings
