### 2022-10-20 08b2788
Initial version of example script
