+++ examples/scripts/denoise_cplx_tv_nlpadmm.py
@@ -94,7 +94,7 @@ hist_tv = solver_tv.itstat_object.history(transpose=True)
 
 
 """
-Denoise with non-linear total variation.
+Denoise with total variation applied to the magnitude of a complex image.
 """
 λ_nltv = 2e-1
 g = λ_nltv * functional.L21Norm()
