### 2022-10-20 fbb05cb
Move iteration stats initialization into separate method
