+++ scico/optimize/_ladmm.py
@@ -113,12 +113,12 @@ class LinearizedADMM:
                 the :class:`.diagnostics.IterationStats` initializer. The
                 dict may also include an additional key "itstat_func"
                 with the corresponding value being a function with two
-                parameters, an integer and a `LinearizedADMM` object,
-                responsible for constructing a tuple ready for insertion
-                into the :class:`.diagnostics.IterationStats` object. If
-                ``None``, default values are used for the dict entries,
-                otherwise the default dict is updated with the dict
-                specified by this parameter.
+                parameters, an integer and a :class:`LinearizedADMM`
+                object, responsible for constructing a tuple ready for
+                insertion into the :class:`.diagnostics.IterationStats`
+                object. If ``None``, default values are used for the dict
+                entries, otherwise the default dict is updated with the
+                dict specified by this parameter.
         """
         self.f: Functional = f
         self.g: Functional = g
@@ -129,6 +129,31 @@ class LinearizedADMM:
         self.maxiter: int = maxiter
         self.timer: Timer = Timer()
 
+        if x0 is None:
+            input_shape = C.input_shape
+            dtype = C.input_dtype
+            x0 = snp.zeros(input_shape, dtype=dtype)
+        self.x = ensure_on_device(x0)
+        self.z, self.z_old = self.z_init(self.x)
+        self.u = self.u_init(self.x)
+
+        self._itstat_init(itstat_options)
+
+    def _itstat_init(self, itstat_options: Optional[dict] = None):
+        """Initialize iteration statistics mechanism.
+
+        Args:
+           itstat_options: A dict of named parameters to be passed to
+                the :class:`.diagnostics.IterationStats` initializer. The
+                dict may also include an additional key "itstat_func"
+                with the corresponding value being a function with two
+                parameters, an integer and a :class:`LinearizedADMM`
+                object, responsible for constructing a tuple ready for
+                insertion into the :class:`.diagnostics.IterationStats`
+                object. If ``None``, default values are used for the dict
+                entries, otherwise the default dict is updated with the
+                dict specified by this parameter.
+        """
         # iteration number and time fields
         itstat_fields = {
             "Iter": "%d",
@@ -136,7 +161,7 @@ class LinearizedADMM:
         }
         itstat_attrib = ["itnum", "timer.elapsed()"]
         # objective function can be evaluated if 'g' function can be evaluated
-        if g.has_eval:
+        if self.g.has_eval:
             itstat_fields.update({"Objective": "%9.3e"})
             itstat_attrib.append("objective()")
         # primal and dual residual fields
@@ -158,14 +183,6 @@ class LinearizedADMM:
             default_itstat_options.update(itstat_options)
         self.itstat_insert_func: Callable = default_itstat_options.pop("itstat_func", None)  # type: ignore
         self.itstat_object = IterationStats(**default_itstat_options)  # type: ignore
-
-        if x0 is None:
-            input_shape = C.input_shape
-            dtype = C.input_dtype
-            x0 = snp.zeros(input_shape, dtype=dtype)
-        self.x = ensure_on_device(x0)
-        self.z, self.z_old = self.z_init(self.x)
-        self.u = self.u_init(self.x)
 
     def objective(
         self,
