+++ scico/optimize/_admm.py
@@ -208,7 +208,7 @@ class ADMM:
         exec("def itstat_func(obj): " + itstat_return, scope)
 
         # determine itstat options and initialize IterationStats object
-        default_itstat_options = {
+        default_itstat_options: dict[str, Union[dict, Callable, bool]] = {
             "fields": itstat_fields,
             "itstat_func": scope["itstat_func"],
             "display": False,
