### 2022-10-20 #4722 @vitorsr
Hi @danielskatz, I will be concluding my review shortly. The project is
spotless. Any and all comments that I would have are mainly curiosities
with reference to the subject itself in my capacity as a signal processing
researcher, not as a reviewer with reference to the project code submitted
to JOSS.

On Thu, Oct 20, 2022 at 2:03 PM Daniel S. Katz ***@***.***>
wrote:

> 👋 @vitorsr <https://github.com/vitorsr> - do you have any update on your
> status?
>
