+++ scico/optimize/_pgm.py
@@ -16,7 +16,6 @@ from typing import Callable, Optional, Union
 import jax
 
 import scico.numpy as snp
-from scico.diagnostics import IterationStats
 from scico.functional import Functional
 from scico.loss import Loss
 from scico.numpy import BlockArray
@@ -24,6 +23,7 @@ from scico.numpy.util import ensure_on_device
 from scico.typing import JaxArray
 from scico.util import Timer
 
+from ._common import itstat_func_and_object
 from ._pgmaux import (
     AdaptiveBBStepSize,
     BBStepSize,
@@ -133,20 +133,9 @@ class PGM:
         itstat_fields.update({"L": "%9.3e", "Residual": "%9.3e"})
         itstat_attrib.extend(["L", "norm_residual()"])
 
-        # dynamically create itstat_func; see https://stackoverflow.com/questions/24733831
-        itstat_return = "return(" + ", ".join(["obj." + attr for attr in itstat_attrib]) + ")"
-        scope: dict[str, Callable] = {}
-        exec("def itstat_func(obj): " + itstat_return, scope)
-
-        default_itstat_options: dict[str, Union[dict, Callable, bool]] = {
-            "fields": itstat_fields,
-            "itstat_func": scope["itstat_func"],
-            "display": False,
-        }
-        if itstat_options:
-            default_itstat_options.update(itstat_options)
-        self.itstat_insert_func: Callable = default_itstat_options.pop("itstat_func")  # type: ignore
-        self.itstat_object = IterationStats(**default_itstat_options)  # type: ignore
+        self.itstat_insert_func, self.itstat_object = itstat_func_and_object(
+            itstat_fields, itstat_attrib, itstat_options
+        )
 
     def objective(self, x: Optional[Union[JaxArray, BlockArray]] = None) -> float:
         r"""Evaluate the objective function :math:`f(\mb{x}) + g(\mb{x})`."""
