### 2022-10-20 #4722 @vitorsr
A small note on my review. I could not easily provision a GPU and I have
had to do it on my personal computer, so my peruse has been bound to
CPU-only testing (as the authors do in their CI).

On Thu, Oct 20, 2022 at 2:03 PM Daniel S. Katz ***@***.***>
wrote:

> 👋 @vitorsr <https://github.com/vitorsr> - do you have any update on your
> status?
>
