### 2022-10-20 1e0c2af
Add least squares solver function
