+++ scico/solver.py
@@ -65,6 +65,7 @@ import numpy as np
 import jax
 import jax.experimental.host_callback as hcb
 
+import scico.linop
 import scico.numpy as snp
 from scico.numpy import BlockArray
 from scico.typing import BlockShape, DType, JaxArray, Shape
@@ -311,7 +312,7 @@ def cg(
 
     Args:
         A: Function implementing linear operator :math:`A`, should be
-            positive definite.
+           positive definite.
         b: Input array :math:`\mb{b}`.
         x0: Initial solution.
         tol: Relative residual stopping tolerance. Convergence occurs
@@ -365,6 +366,72 @@ def cg(
         return (x, {"num_iter": ii, "rel_res": snp.sqrt(num).real / bn})
     else:
         return x
+
+
+def lstsq(
+    A: Callable,
+    b: JaxArray,
+    x0: Optional[JaxArray] = None,
+    tol: float = 1e-5,
+    atol: float = 0.0,
+    maxiter: int = 1000,
+    info: bool = False,
+    M: Optional[Callable] = None,
+) -> Tuple[JaxArray, dict]:
+    r"""Conjugate Gradient solver.
+
+    Solve the least squares problem
+
+    .. math::
+        \argmin_{\mb{x}} \; (1/2) \norm{ A \mb{x} - \mb{b}) }_2^2 \;,
+
+    where :math:`A` is a linear operator and :math:`\mb{b}` is a vector.
+    The problem is solved using :func:`cg`.
+
+    Args:
+        A: Callable implementing linear operator :math:`A`.
+        b: Input array :math:`\mb{b}`.
+        x0: Initial solution. If `A` is a :class:`.LinearOperator`, this
+          parameter need to be specified, and defaults to a zero array.
+          Otherwise, it is required.
+        tol: Relative residual stopping tolerance. Convergence occurs
+           when `norm(residual) <= max(tol * norm(b), atol)`.
+        atol: Absolute residual stopping tolerance. Convergence occurs
+           when `norm(residual) <= max(tol * norm(b), atol)`.
+        maxiter: Maximum iterations. Default: 1000.
+        info: If ``True`` return a tuple consting of the solution array
+           and a dictionary containing diagnostic information, otherwise
+           just return the solution.
+        M: Preconditioner for `A`. The preconditioner should approximate
+           the inverse of `A`. The default, ``None``, uses no
+           preconditioner.
+
+    Returns:
+        tuple: A tuple (x, info) containing:
+
+            - **x** : Solution array.
+            - **info**: Dictionary containing diagnostic information.
+    """
+    if x0 is None:
+        if isinstance(A, scico.linop.LinearOperator):
+            x0 = snp.zeros(A.input_shape, b.dtype)
+        else:
+            raise ValueError("Parameter x0 must be specified if A is not a LinearOperator")
+
+    if isinstance(A, scico.linop.LinearOperator):
+        Aop = A
+    else:
+        Aop = scico.linop.LinearOperator(
+            input_shape=x0.shape,
+            output_shape=b.shape,
+            eval_fn=A,
+            input_dtype=b.dtype,
+            output_dtype=b.dtype,
+        )
+
+    ATA = Aop.T @ Aop
+    ATb = Aop.T @ b
+    return cg(ATA, ATb, x0=x0, tol=tol, atol=atol, maxiter=maxiter, info=info, M=M)
 
 
 def bisect(
