### 2022-10-20 f48f844
Address mypy error
