### 2022-10-20 bb082b3
Address mypy error
