### 2022-10-21 05b7e86
Fix use of local copy of mathjax 3
