### 2022-10-21 4dca212
Fix imgmath option for offline docs build
