### 2022-10-22 #4722 @lucaferranti
Dear @danielskatz and  @bwohlberg 👋 ,

First, I would like to start by sincerely apologizing for the huuuge delay in carrying my review. I have been a lot of deadlines and being the software pretty big it took a while to go through the code and examples. Again, I am very sorry for the delay, but finally I got there.

So, in general I would say the software and the paper have an exceptionally high quality and I definitely recommend for publication.

The software is very well documented and I particularly like how it supports  both binder and colab to allow users to start playing with the examples. I have also run the examples locally and they seem to work appropriately.

The paper is in general well written. I only have one comment:

- [ ] currently there is no comparison with other software in the fields. The authors mention in the statement of need that there are other packages overlapping with SCICO functionalities and a few also support execution on GPU, but they don't mention by name any other library. I would encourage the authors to expand and cite relevant software in the field; ideally also comparing SCICO to them in more details, e.g. in table form.
