### 2022-10-22 #4722 @lucaferranti
Regarding the references, the list of DOIs above seems quite longer than the list of references in the paper, indeed the `references.bib` seem to have several references which are not cited in the paper. Could you clarify on their role?

@danielskatz I presume the references file should only contain the references which are actually used in the paper? Can you comment from an editor perspective on this?
