### 2022-10-22 #4722 @danielskatz
@lucaferranti - regarding the references, our reference checker checks everything in the bib file, so references that are not used in the paper don't need to be there, but as long as they aren't causing problems in the check, they won't cause any problems anywhere else either
