### 2022-10-22 #4722 @bwohlberg
Thank you @lucaferranti and @danielskatz. The requested comparison with other software can be found in section _Related Packages_ [on this docs page](https://scico.readthedocs.io/en/stable/advantages.html) (I suspect that @lucaferranti did not notice it because it was added relatively recently). We have not intentionally included any references that are not cited at least once (some may only be cited in a single example script), but it is possible that such references have inadvertently been created during changes to the docs.
