### 2022-10-23 e775cb3
Move support for optional x0 into cg function
