### 2022-10-23 e49ad4d
Resolve mypy error
