### 2022-10-24 edc00ed
Resolve issue #291 (#355)
* Move iteration stats initialization into separate method

* Add typing

* Clean up

* Address mypy error
