+++ scico/optimize/_ladmm.py
@@ -14,7 +14,6 @@ from __future__ import annotations
 from typing import Callable, List, Optional, Tuple, Union
 
 import scico.numpy as snp
-from scico.diagnostics import IterationStats
 from scico.functional import Functional
 from scico.linop import LinearOperator
 from scico.numpy import BlockArray
@@ -22,6 +21,8 @@ from scico.numpy.linalg import norm
 from scico.numpy.util import ensure_on_device
 from scico.typing import JaxArray
 from scico.util import Timer
+
+from ._common import itstat_func_and_object
 
 
 class LinearizedADMM:
@@ -113,12 +114,12 @@ class LinearizedADMM:
                 the :class:`.diagnostics.IterationStats` initializer. The
                 dict may also include an additional key "itstat_func"
                 with the corresponding value being a function with two
-                parameters, an integer and a `LinearizedADMM` object,
-                responsible for constructing a tuple ready for insertion
-                into the :class:`.diagnostics.IterationStats` object. If
-                ``None``, default values are used for the dict entries,
-                otherwise the default dict is updated with the dict
-                specified by this parameter.
+                parameters, an integer and a :class:`LinearizedADMM`
+                object, responsible for constructing a tuple ready for
+                insertion into the :class:`.diagnostics.IterationStats`
+                object. If ``None``, default values are used for the dict
+                entries, otherwise the default dict is updated with the
+                dict specified by this parameter.
         """
         self.f: Functional = f
         self.g: Functional = g
@@ -129,36 +130,6 @@ class LinearizedADMM:
         self.maxiter: int = maxiter
         self.timer: Timer = Timer()
 
-        # iteration number and time fields
-        itstat_fields = {
-            "Iter": "%d",
-            "Time": "%8.2e",
-        }
-        itstat_attrib = ["itnum", "timer.elapsed()"]
-        # objective function can be evaluated if 'g' function can be evaluated
-        if g.has_eval:
-            itstat_fields.update({"Objective": "%9.3e"})
-            itstat_attrib.append("objective()")
-        # primal and dual residual fields
-        itstat_fields.update({"Prml Rsdl": "%9.3e", "Dual Rsdl": "%9.3e"})
-        itstat_attrib.extend(["norm_primal_residual()", "norm_dual_residual()"])
-
-        # dynamically create itstat_func; see https://stackoverflow.com/questions/24733831
-        itstat_return = "return(" + ", ".join(["obj." + attr for attr in itstat_attrib]) + ")"
-        scope: dict[str, Callable] = {}
-        exec("def itstat_func(obj): " + itstat_return, scope)
-
-        # determine itstat options and initialize IterationStats object
-        default_itstat_options = {
-            "fields": itstat_fields,
-            "itstat_func": scope["itstat_func"],
-            "display": False,
-        }
-        if itstat_options:
-            default_itstat_options.update(itstat_options)
-        self.itstat_insert_func: Callable = default_itstat_options.pop("itstat_func", None)  # type: ignore
-        self.itstat_object = IterationStats(**default_itstat_options)  # type: ignore
-
         if x0 is None:
             input_shape = C.input_shape
             dtype = C.input_dtype
@@ -166,6 +137,41 @@ class LinearizedADMM:
         self.x = ensure_on_device(x0)
         self.z, self.z_old = self.z_init(self.x)
         self.u = self.u_init(self.x)
+
+        self._itstat_init(itstat_options)
+
+    def _itstat_init(self, itstat_options: Optional[dict] = None):
+        """Initialize iteration statistics mechanism.
+
+        Args:
+           itstat_options: A dict of named parameters to be passed to
+                the :class:`.diagnostics.IterationStats` initializer. The
[…]
