### 2022-10-24 0970bdf
remove subscript from indicator
