+++ scico/functional/_indicator.py
@@ -26,7 +26,7 @@ class NonNegativeIndicator(Functional):
     inf otherwise
 
     .. math::
-        I_{>= 0}(\mb{x}) = \begin{cases}
+        I(\mb{x}) = \begin{cases}
         0  & \text{ if } x_i \geq 0 \; \forall i \\
         \infty  & \text{ otherwise} \;.
         \end{cases}
@@ -48,7 +48,7 @@ class NonNegativeIndicator(Functional):
         r"""The scaled proximal operator of the non-negative indicator.
 
         Evaluate the scaled proximal operator of the indicator over
-        the non-negative orthant, :math:`I_{>= 0}`,
+        the non-negative orthant, :math:`I`,
 
         .. math::
             [\mathrm{prox}_{\lambda I_{>=0}}(\mb{v})]_i =
@@ -72,7 +72,7 @@ class L2BallIndicator(Functional):
     Indicator function for :math:`\ell_2` ball of given radius, :math:`r`
 
     .. math::
-        I_r(\mb{x}) = \begin{cases}
+        I(\mb{x}) = \begin{cases}
         0  & \text{ if } \norm{\mb{x}}_2 \leq r \\
         \infty  & \text{ otherwise} \;.
         \end{cases}
@@ -104,10 +104,10 @@ class L2BallIndicator(Functional):
         r"""The scaled proximal operator of the :math:`\ell_2` ball indicator.
         a :math:`\ell_2` ball
 
-        Evaluate the scaled proximal operator of the indicator, :math:`I_r`,
+        Evaluate the scaled proximal operator of the indicator, :math:`I`,
         of the :math:`\ell_2` ball with radius :math:`r`
 
         .. math::
-            \mathrm{prox}_{\lambda I_r}(\mb{v}) = r \frac{\mb{v}}{\norm{\mb{v}}_2}\;.
+            \mathrm{prox}_{\lambda I}(\mb{v}) = r \frac{\mb{v}}{\norm{\mb{v}}_2}\;.
         """
         return self.radius * v / norm(v)
