### 2022-10-24 ed1c82f
Resolve issue #240 (#354)
* Add docs and implementation of parameter that had no effect

* Add least squares solver function

* Docs fix

* Move support for optional x0 into cg function

* Resolve mypy error

* Typo fix
