### 2022-10-24 7498b95
Indicator documentation notation (#353)
* Indicator documentation notation

Make class docstrings agree with prox docstrings

* remove subscript from indicator

* fix missing subscript

Co-authored-by: Brendt Wohlberg <bwohlberg@users.noreply.github.com>
