### 2022-10-25 #4722 @DanNixon
Hi all, apologies from me also that my review has taken longer than I said it would, a few things came up that meant I only had the odd bit of time in the evenings to look at the software itself.

Anyway, I am happy with what I have seen and don't have any significant comments beyond what has already been said. The project is very well put together with good documentation and extensive examples/use cases.
