### 2022-10-25 #4722 @vitorsr
All done. My only concern is that the JAX dependency seems a bit fragile - is there any plan to "future-proof" this requirement?
