### 2022-10-25 #4722 @bwohlberg
> All done. My only concern is that the JAX dependency seems a bit fragile - is there any plan to "future-proof" this requirement?

We have experienced some issues with changes in `jaxlib` and `jax` breaking parts of our code, but a recent re-write of the `BlockArray` implementation has made that much less frequent. The project `requirements.txt` file includes both minimum and maximum `jaxlib` / `jax` version numbers to ensure that the user installs versions of these dependencies that are known to function correctly with `scico`, and we've set up a weekly job in the project CI to check whether there are any compatibility issues with any new  `jaxlib` / `jax` releases. We're open to suggestions, though, if there's anything else that could be done that we've overlooked.
