### 2022-10-25 #4722 @vitorsr
Yes. The submission has no acceptance blockers and meets all acceptance
criteria.

On Tue, Oct 25, 2022 at 10:35 AM Daniel S. Katz ***@***.***>
wrote:

> 👋 @vitorsr <https://github.com/vitorsr>, @DanNixon
> <https://github.com/DanNixon>, @lucaferranti
> <https://github.com/lucaferranti> - you've all now checked off all your
> review items, which I presume means that you think this can be accepted.
> Please confirm this is the case.
>
