### 2022-10-25 #4722 @danielskatz
👋 @bwohlberg - at this point could you:

I can then move forward with the next steps, which will include me proofreading the paper, perhaps suggesting changes, then accepting and publishing the submission.
