### 2022-10-25 #4722 @bwohlberg
Thank you @vitorsr, @DanNixon, and @lucaferranti for your reviews.

@danielskatz: We released version 0.0.3 about a month ago (well after the review had started) and have only made minor changes since then. Would it be OK to use that tagged release for the process outlined above? Also, any objections to some minor changes (e.g. correcting a reference, updating a description of the repo) being made to the paper itself at this stage?
