### 2022-10-26 #4722 @danielskatz
Both of those are fine.
