+++ requirements.txt
@@ -5,7 +5,7 @@ imageio>=2.17
 matplotlib
 jaxlib>=0.3.2,<=0.3.15
 jax>=0.3.2,<=0.3.17
-flax>=0.5.3
+flax>=0.4.2,<=0.5.3
 bm3d>=3.0.9
 svmbir>=0.3.0
 pyabel>=0.8.5
