+++ docs/joss-paper/paper.bib
@@ -151,10 +151,10 @@
                   T. Buzzard and Brendt Wohlberg},
   title =	 {Plug-and-Play Methods for Integrating Physical and
                   Learned Models in Computational Imaging},
-  journal =      {IEEE Signal Processing Magazine},
+  journal =      {{\normalfont To {a}ppear in} IEEE Signal Processing Magazine},
   year =	 2022,
+  doi =          {10.48550/arXiv.2203.17061},
   eprint =	 {arXiv:2203.17061},
-  note =         {To appear.}
 }
 
 @Software {bradbury-2018-jax,
