### 2022-10-27 2e9e46b
Remove config for mathjax2
