+++ requirements.txt
@@ -3,8 +3,8 @@ scipy>=1.6.0
 tifffile
 imageio>=2.17
 matplotlib
-jaxlib>=0.3.0,<=0.3.15
-jax>=0.3.0,<=0.3.17
+jaxlib>=0.3.0,<=0.3.22
+jax>=0.3.0,<=0.3.23
 flax>=0.4.0
 bm3d>=3.0.9
 svmbir>=0.3.0
