### 2022-10-27 ea90343
Fix typo and update description of pending ML components
