### 2022-10-28 40f043c
Add tests for matrix input shapes
