### 2022-10-28 #4722 @danielskatz
@bwohlberg - I just have one small change from my proofreading: https://github.com/lanl/scico/pull/361   can you please merge this? Then we can continue the process.
