### 2022-10-28 #4722 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
